Explain a random forest

Explains a random forest in a html document using plots created by randomForestExplainer

explain_forest(forest, interactions = FALSE, data = NULL,
  vars = NULL, no_of_pred_plots = 3, pred_grid = 100,
  measures = NULL)

Arguments

forest
A randomForest object created with the option localImp = TRUE

interactions
Logical value: should variable interactions be considered (this may be time-consuming)

data
The data frame on which forest was trained - necessary if interactions = TRUE

vars
A character vector with variables with respect to which interactions will be considered if NULL then they will be selected using the important_variables() function

no_of_pred_plots
The number of most frequent interactions of numeric variables to plot predictions for

pred_grid
The number of points on the grid of plot_predict_interaction (decrease in case memory problems)

measures
A character vector specifying the importance measures to be used for plotting ggpairs

Value

A html document in your working directory

Examples

# \dontrun{
forest <- randomForest::randomForest(Species ~ ., data = iris, localImp = TRUE)
explain_forest(forest, interactions = TRUE)

#> 
#> 
#> processing file: Explain_forest_template.Rmd

#>   |                                                                         |                                                                 |   0%  |                                                                         |...                                                              |   5%
#>    inline R code fragments
#> 
#>   |                                                                         |......                                                           |  10%
#> label: setup (with options) 
#> List of 1
#>  $ include: logi FALSE
#> 
#>   |                                                                         |..........                                                       |  15%
#>   ordinary text without R code
#> 
#>   |                                                                         |.............                                                    |  20%
#> label: unnamed-chunk-1
#>   |                                                                         |................                                                 |  25%
#>   ordinary text without R code
#> 
#>   |                                                                         |....................                                             |  30%
#> label: unnamed-chunk-2
#>   |                                                                         |.......................                                          |  35%
#>   ordinary text without R code
#> 
#>   |                                                                         |..........................                                       |  40%
#> label: unnamed-chunk-3

#>   |                                                                         |.............................                                    |  45%
#>   ordinary text without R code
#> 
#>   |                                                                         |................................                                 |  50%
#> label: unnamed-chunk-4
#>   |                                                                         |....................................                             |  55%
#>   ordinary text without R code
#> 
#>   |                                                                         |.......................................                          |  60%
#> label: unnamed-chunk-5

#>   |                                                                         |..........................................                       |  65%
#>    inline R code fragments
#> 
#> [1] accuracy_decrease and gini_decrease
#>   |                                                                         |..............................................                   |  70%
#> label: unnamed-chunk-6

#> Warning: Using alpha for a discrete variable is not advised.

#>   |                                                                         |.................................................                |  75%
#>    inline R code fragments
#> 
#> [1] mean_min_depth, accuracy_decrease, gini_decrease, no_of_nodes, times_a_root
#>   |                                                                         |....................................................             |  80%
#> label: unnamed-chunk-7 (with options) 
#> List of 2
#>  $ fig.width : num 10
#>  $ fig.height: num 9
#> 

#>   |                                                                         |.......................................................          |  85%
#>   ordinary text without R code
#> 
#>   |                                                                         |..........................................................       |  90%
#> label: unnamed-chunk-8 (with options) 
#> List of 2
#>  $ fig.width : num 10
#>  $ fig.height: num 9
#> 

#> Warning: span too small.   fewer data values than degrees of freedom.

#> Warning: pseudoinverse used at 0.985

#> Warning: neighborhood radius 2.015

#> Warning: reciprocal condition number  0

#> Warning: There are other near singularities as well. 4.0602

#> Warning: span too small.   fewer data values than degrees of freedom.

#> Warning: pseudoinverse used at 0.985

#> Warning: neighborhood radius 2.015

#> Warning: reciprocal condition number  0

#> Warning: There are other near singularities as well. 4.0602

#> Warning: span too small.   fewer data values than degrees of freedom.

#> Warning: pseudoinverse used at 0.985

#> Warning: neighborhood radius 2.015

#> Warning: reciprocal condition number  0

#> Warning: There are other near singularities as well. 4.0602

#> Warning: span too small.   fewer data values than degrees of freedom.

#> Warning: pseudoinverse used at 0.985

#> Warning: neighborhood radius 2.015

#> Warning: reciprocal condition number  0

#> Warning: There are other near singularities as well. 4.0602

#> Warning: span too small.   fewer data values than degrees of freedom.

#> Warning: pseudoinverse used at 0.985

#> Warning: neighborhood radius 2.015

#> Warning: reciprocal condition number  0

#> Warning: There are other near singularities as well. 4.0602

#> Warning: span too small.   fewer data values than degrees of freedom.

#> Warning: pseudoinverse used at 0.985

#> Warning: neighborhood radius 2.015

#> Warning: reciprocal condition number  0

#> Warning: There are other near singularities as well. 4.0602

#> Warning: span too small.   fewer data values than degrees of freedom.

#> Warning: pseudoinverse used at 0.985

#> Warning: neighborhood radius 2.015

#> Warning: reciprocal condition number  0

#> Warning: There are other near singularities as well. 4.0602

#> Warning: span too small.   fewer data values than degrees of freedom.

#> Warning: pseudoinverse used at 0.985

#> Warning: neighborhood radius 2.015

#> Warning: reciprocal condition number  0

#> Warning: There are other near singularities as well. 4.0602

#> Warning: span too small.   fewer data values than degrees of freedom.

#> Warning: pseudoinverse used at 0.985

#> Warning: neighborhood radius 2.015

#> Warning: reciprocal condition number  0

#> Warning: There are other near singularities as well. 4.0602

#> Warning: span too small.   fewer data values than degrees of freedom.

#> Warning: pseudoinverse used at 0.985

#> Warning: neighborhood radius 2.015

#> Warning: reciprocal condition number  0

#> Warning: There are other near singularities as well. 4.0602

#> Warning: span too small.   fewer data values than degrees of freedom.

#> Warning: pseudoinverse used at 0.985

#> Warning: neighborhood radius 2.015

#> Warning: reciprocal condition number  0

#> Warning: There are other near singularities as well. 4.0602

#> Warning: span too small.   fewer data values than degrees of freedom.

#> Warning: pseudoinverse used at 0.985

#> Warning: neighborhood radius 2.015

#> Warning: reciprocal condition number  0

#> Warning: There are other near singularities as well. 4.0602

#> Warning: span too small.   fewer data values than degrees of freedom.

#> Warning: pseudoinverse used at 0.985

#> Warning: neighborhood radius 2.015

#> Warning: reciprocal condition number  0

#> Warning: There are other near singularities as well. 4.0602

#> Warning: span too small.   fewer data values than degrees of freedom.

#> Warning: pseudoinverse used at 0.985

#> Warning: neighborhood radius 2.015

#> Warning: reciprocal condition number  0

#> Warning: There are other near singularities as well. 4.0602

#> Warning: span too small.   fewer data values than degrees of freedom.

#> Warning: pseudoinverse used at 0.985

#> Warning: neighborhood radius 2.015

#> Warning: reciprocal condition number  0

#> Warning: There are other near singularities as well. 4.0602

#> Warning: span too small.   fewer data values than degrees of freedom.

#> Warning: pseudoinverse used at 0.985

#> Warning: neighborhood radius 2.015

#> Warning: reciprocal condition number  0

#> Warning: There are other near singularities as well. 4.0602

#> Warning: span too small.   fewer data values than degrees of freedom.

#> Warning: pseudoinverse used at 0.985

#> Warning: neighborhood radius 2.015

#> Warning: reciprocal condition number  0

#> Warning: There are other near singularities as well. 4.0602

#> Warning: span too small.   fewer data values than degrees of freedom.

#> Warning: pseudoinverse used at 0.985

#> Warning: neighborhood radius 2.015

#> Warning: reciprocal condition number  0

#> Warning: There are other near singularities as well. 4.0602

#> Warning: span too small.   fewer data values than degrees of freedom.

#> Warning: pseudoinverse used at 0.985

#> Warning: neighborhood radius 2.015

#> Warning: reciprocal condition number  0

#> Warning: There are other near singularities as well. 4.0602

#> Warning: span too small.   fewer data values than degrees of freedom.

#> Warning: pseudoinverse used at 0.985

#> Warning: neighborhood radius 2.015

#> Warning: reciprocal condition number  0

#> Warning: There are other near singularities as well. 4.0602

#>   |                                                                         |..............................................................   |  95%
#>   ordinary text without R code
#> 
#>   |                                                                         |.................................................................| 100%
#> label: conditional_print (with options) 
#> List of 2
#>  $ child: chr "Explain_forest_template_interactions.Rmd"
#>  $ eval : symbol interactions
#> 

#> 
#> 
#> processing file: ./Explain_forest_template_interactions.Rmd

#>   |                                                                         |                                                                 |   0%  |                                                                         |.........                                                        |  14%
#>    inline R code fragments
#> 
#> [1] Petal.Width, Petal.Length, Sepal.Length, Sepal.Width
#>   |                                                                         |...................                                              |  29%
#> label: unnamed-chunk-9

#>   |                                                                         |............................                                     |  43%
#>   ordinary text without R code
#> 
#>   |                                                                         |.....................................                            |  57%
#> label: unnamed-chunk-10
#>   |                                                                         |..............................................                   |  71%
#>    inline R code fragments
#> 
#>   |                                                                         |........................................................         |  86%
#> label: unnamed-chunk-11
#>   |                                                                         |.................................................................| 100%
#>   ordinary text without R code
#> 
#> 
#> 

#> output file: Explain_forest_template.knit.md

#> /Applications/RStudio.app/Contents/MacOS/pandoc/pandoc +RTS -K512m -RTS Explain_forest_template.utf8.md --to html4 --from markdown+autolink_bare_uris+ascii_identifiers+tex_math_single_backslash+smart --output /Users/pbiecek/GitHub/randomForestExplainer/docs/reference/Your_forest_explained.html --email-obfuscation none --self-contained --standalone --section-divs --table-of-contents --toc-depth 3 --variable toc_float=1 --variable toc_selectors=h1,h2,h3 --variable toc_collapsed=1 --variable toc_smooth_scroll=1 --variable toc_print=1 --template /Library/Frameworks/R.framework/Versions/3.5/Resources/library/rmarkdown/rmd/h/default.html --no-highlight --variable highlightjs=1 --variable 'theme:bootstrap' --include-in-header /var/folders/_l/jllqh32s3llbxmtrh2431vpr0000gn/T//RtmpWFwgg7/rmarkdown-str1384b6867d71.html --mathjax --variable 'mathjax-url:https://mathjax.rstudio.com/latest/MathJax.js?config=TeX-AMS-MML_HTMLorMML' 

#> 
#> Output created: Your_forest_explained.html

# }