Calculate minimal depth distribution of a random forest

Get minimal depth values for all trees in a random forest

min_depth_distribution(forest)

Arguments

forest
A randomForest or ranger object

Value

A data frame with the value of minimal depth for every variable in every tree

Examples

min_depth_distribution(randomForest::randomForest(Species ~ ., data = iris, ntree = 100))

#>     tree     variable minimal_depth
#> 1      1 Petal.Length             1
#> 2      1  Petal.Width             0
#> 3      1 Sepal.Length             4
#> 4      1  Sepal.Width             2
#> 5      2 Petal.Length             0
#> 6      2  Petal.Width             2
#> 7      2 Sepal.Length             2
#> 8      3 Petal.Length             2
#> 9      3  Petal.Width             0
#> 10     3  Sepal.Width             3
#> 11     4 Petal.Length             0
#> 12     4  Petal.Width             1
#> 13     4 Sepal.Length             2
#> 14     4  Sepal.Width             3
#> 15     5 Petal.Length             2
#> 16     5  Petal.Width             0
#> 17     5 Sepal.Length             1
#> 18     5  Sepal.Width             2
#> 19     6 Petal.Length             0
#> 20     6  Petal.Width             1
#> 21     6 Sepal.Length             3
#> 22     6  Sepal.Width             3
#> 23     7 Petal.Length             0
#> 24     7  Petal.Width             1
#> 25     7 Sepal.Length             2
#> 26     7  Sepal.Width             4
#> 27     8 Petal.Length             0
#> 28     8  Petal.Width             2
#> 29     8  Sepal.Width             3
#> 30     9 Petal.Length             5
#> 31     9  Petal.Width             0
#> 32     9 Sepal.Length             2
#> 33     9  Sepal.Width             3
#> 34    10 Petal.Length             0
#> 35    10  Petal.Width             2
#> 36    10 Sepal.Length             4
#> 37    10  Sepal.Width             3
#> 38    11 Petal.Length             1
#> 39    11  Petal.Width             0
#> 40    11 Sepal.Length             4
#> 41    11  Sepal.Width             2
#> 42    12 Petal.Length             2
#> 43    12  Petal.Width             0
#> 44    12 Sepal.Length             1
#> 45    12  Sepal.Width             4
#> 46    13 Petal.Length             2
#> 47    13  Petal.Width             0
#> 48    13 Sepal.Length             3
#> 49    13  Sepal.Width             2
#> 50    14 Petal.Length             0
#> 51    14  Petal.Width             1
#> 52    14  Sepal.Width             3
#> 53    15 Petal.Length             0
#> 54    15  Petal.Width             4
#> 55    15 Sepal.Length             4
#> 56    15  Sepal.Width             2
#> 57    16 Petal.Length             1
#> 58    16  Petal.Width             1
#> 59    16 Sepal.Length             0
#> 60    16  Sepal.Width             2
#> 61    17 Petal.Length             2
#> 62    17  Petal.Width             0
#> 63    18 Petal.Length             0
#> 64    18  Petal.Width             1
#> 65    18 Sepal.Length             3
#> 66    18  Sepal.Width             3
#> 67    19 Petal.Length             2
#> 68    19  Petal.Width             0
#> 69    20 Petal.Length             4
#> 70    20  Petal.Width             0
#> 71    20 Sepal.Length             2
#> 72    20  Sepal.Width             5
#> 73    21 Petal.Length             1
#> 74    21  Petal.Width             0
#> 75    21 Sepal.Length             3
#> 76    22 Petal.Length             0
#> 77    22  Petal.Width             3
#> 78    22 Sepal.Length             3
#> 79    22  Sepal.Width             3
#> 80    23 Petal.Length             0
#> 81    23  Petal.Width             1
#> 82    24 Petal.Length             3
#> 83    24  Petal.Width             0
#> 84    24 Sepal.Length             2
#> 85    24  Sepal.Width             4
#> 86    25 Petal.Length             0
#> 87    25  Petal.Width             1
#> 88    25 Sepal.Length             2
#> 89    25  Sepal.Width             3
#> 90    26 Petal.Length             1
#> 91    26  Petal.Width             2
#> 92    26 Sepal.Length             0
#> 93    26  Sepal.Width             1
#> 94    27 Petal.Length             1
#> 95    27  Petal.Width             0
#> 96    27 Sepal.Length             3
#> 97    27  Sepal.Width             4
#> 98    28 Petal.Length             0
#> 99    28  Petal.Width             2
#> 100   28  Sepal.Width             3
#> 101   29 Petal.Length             1
#> 102   29  Petal.Width             0
#> 103   29 Sepal.Length             2
#> 104   29  Sepal.Width             3
#> 105   30 Petal.Length             0
#> 106   30  Petal.Width             2
#> 107   30  Sepal.Width             3
#> 108   31 Petal.Length             2
#> 109   31  Petal.Width             0
#> 110   31 Sepal.Length             1
#> 111   32 Petal.Length             0
#> 112   32  Petal.Width             1
#> 113   32 Sepal.Length             3
#> 114   32  Sepal.Width             3
#> 115   33 Petal.Length             1
#> 116   33  Petal.Width             0
#> 117   33 Sepal.Length             2
#> 118   34 Petal.Length             0
#> 119   34  Petal.Width             2
#> 120   34 Sepal.Length             1
#> 121   34  Sepal.Width             3
#> 122   35 Petal.Length             2
#> 123   35  Petal.Width             1
#> 124   35 Sepal.Length             0
#> 125   35  Sepal.Width             3
#> 126   36 Petal.Length             2
#> 127   36  Petal.Width             0
#> 128   36 Sepal.Length             4
#> 129   36  Sepal.Width             2
#> 130   37 Petal.Length             1
#> 131   37  Petal.Width             0
#> 132   37 Sepal.Length             2
#> 133   37  Sepal.Width             6
#> 134   38 Petal.Length             2
#> 135   38  Petal.Width             0
#> 136   38  Sepal.Width             3
#> 137   39 Petal.Length             0
#> 138   39  Petal.Width             1
#> 139   39 Sepal.Length             3
#> 140   39  Sepal.Width             3
#> 141   40 Petal.Length             0
#> 142   40  Petal.Width             2
#> 143   40  Sepal.Width             2
#> 144   41 Petal.Length             1
#> 145   41  Petal.Width             2
#> 146   41 Sepal.Length             0
#> 147   41  Sepal.Width             5
#> 148   42 Petal.Length             2
#> 149   42  Petal.Width             0
#> 150   42 Sepal.Length             1
#> 151   43 Petal.Length             1
#> 152   43  Petal.Width             0
#> 153   43 Sepal.Length             3
#> 154   43  Sepal.Width             3
#> 155   44 Petal.Length             0
#> 156   44  Petal.Width             2
#> 157   44 Sepal.Length             2
#> 158   44  Sepal.Width             4
#> 159   45 Petal.Length             0
#> 160   45  Petal.Width             1
#> 161   46 Petal.Length             1
#> 162   46  Petal.Width             0
#> 163   46 Sepal.Length             1
#> 164   47 Petal.Length             0
#> 165   47 Sepal.Length             2
#> 166   47  Sepal.Width             3
#> 167   48 Petal.Length             0
#> 168   48  Petal.Width             3
#> 169   48 Sepal.Length             1
#> 170   48  Sepal.Width             4
#> 171   49 Petal.Length             2
#> 172   49  Petal.Width             0
#> 173   49 Sepal.Length             1
#> 174   49  Sepal.Width             2
#> 175   50 Petal.Length             1
#> 176   50  Petal.Width             0
#> 177   51 Petal.Length             0
#> 178   51  Petal.Width             2
#> 179   51 Sepal.Length             1
#> 180   51  Sepal.Width             3
#> 181   52 Petal.Length             0
#> 182   52  Petal.Width             3
#> 183   52 Sepal.Length             1
#> 184   52  Sepal.Width             2
#> 185   53 Petal.Length             1
#> 186   53  Petal.Width             0
#> 187   53  Sepal.Width             4
#> 188   54 Petal.Length             0
#> 189   54  Petal.Width             1
#> 190   54 Sepal.Length             3
#> 191   54  Sepal.Width             3
#> 192   55 Petal.Length             0
#> 193   55  Petal.Width             1
#> 194   55 Sepal.Length             2
#> 195   56 Petal.Length             0
#> 196   56  Petal.Width             1
#> 197   56 Sepal.Length             3
#> 198   56  Sepal.Width             4
#> 199   57 Petal.Length             1
#> 200   57  Petal.Width             1
#> 201   57 Sepal.Length             0
#> 202   57  Sepal.Width             3
#> 203   58 Petal.Length             1
#> 204   58  Petal.Width             1
#> 205   58 Sepal.Length             0
#> 206   59 Petal.Length             0
#> 207   59  Petal.Width             1
#> 208   59 Sepal.Length             4
#> 209   60 Petal.Length             2
#> 210   60  Petal.Width             0
#> 211   60  Sepal.Width             5
#> 212   61 Petal.Length             1
#> 213   61  Petal.Width             3
#> 214   61 Sepal.Length             1
#> 215   61  Sepal.Width             0
#> 216   62 Petal.Length             2
#> 217   62  Petal.Width             0
#> 218   62 Sepal.Length             3
#> 219   62  Sepal.Width             3
#> 220   63 Petal.Length             1
#> 221   63  Petal.Width             0
#> 222   63 Sepal.Length             2
#> 223   63  Sepal.Width             3
#> 224   64 Petal.Length             3
#> 225   64  Petal.Width             0
#> 226   64 Sepal.Length             1
#> 227   64  Sepal.Width             3
#> 228   65 Petal.Length             1
#> 229   65  Petal.Width             1
#> 230   65 Sepal.Length             0
#> 231   65  Sepal.Width             3
#> 232   66 Petal.Length             2
#> 233   66  Petal.Width             0
#> 234   66  Sepal.Width             5
#> 235   67 Petal.Length             0
#> 236   67 Sepal.Length             2
#> 237   68 Petal.Length             0
#> 238   68  Petal.Width             2
#> 239   68 Sepal.Length             3
#> 240   69 Petal.Length             0
#> 241   69  Petal.Width             1
#> 242   69 Sepal.Length             3
#> 243   69  Sepal.Width             4
#> 244   70 Petal.Length             0
#> 245   70  Petal.Width             2
#> 246   70  Sepal.Width             4
#> 247   71 Petal.Length             1
#> 248   71  Petal.Width             1
#> 249   71 Sepal.Length             0
#> 250   72 Petal.Length             1
#> 251   72  Petal.Width             0
#> 252   72 Sepal.Length             2
#> 253   72  Sepal.Width             3
#> 254   73 Petal.Length             0
#> 255   73  Petal.Width             1
#> 256   73 Sepal.Length             2
#> 257   73  Sepal.Width             4
#> 258   74 Petal.Length             0
#> 259   74  Petal.Width             1
#> 260   74 Sepal.Length             2
#> 261   75 Petal.Length             1
#> 262   75  Petal.Width             0
#> 263   75 Sepal.Length             3
#> 264   76 Petal.Length             1
#> 265   76  Petal.Width             0
#> 266   76  Sepal.Width             4
#> 267   77 Petal.Length             0
#> 268   77  Petal.Width             2
#> 269   77  Sepal.Width             3
#> 270   78 Petal.Length             3
#> 271   78  Petal.Width             1
#> 272   78 Sepal.Length             0
#> 273   78  Sepal.Width             1
#> 274   79 Petal.Length             3
#> 275   79  Petal.Width             0
#> 276   79 Sepal.Length             1
#> 277   79  Sepal.Width             4
#> 278   80 Petal.Length             1
#> 279   80  Petal.Width             1
#> 280   80 Sepal.Length             0
#> 281   80  Sepal.Width             4
#> 282   81 Petal.Length             2
#> 283   81  Petal.Width             0
#> 284   82 Petal.Length             2
#> 285   82  Petal.Width             0
#> 286   82 Sepal.Length             4
#> 287   82  Sepal.Width             3
#> 288   83 Petal.Length             1
#> 289   83  Petal.Width             1
#> 290   83 Sepal.Length             0
#> 291   84 Petal.Length             2
#> 292   84  Petal.Width             0
#> 293   84  Sepal.Width             2
#> 294   85 Petal.Length             0
#> 295   85  Petal.Width             2
#> 296   85 Sepal.Length             2
#> 297   85  Sepal.Width             3
#> 298   86 Petal.Length             1
#> 299   86  Petal.Width             2
#> 300   86 Sepal.Length             0
#> 301   86  Sepal.Width             1
#> 302   87 Petal.Length             0
#> 303   87  Petal.Width             3
#> 304   87  Sepal.Width             2
#> 305   88 Petal.Length             0
#> 306   88  Petal.Width             3
#> 307   88 Sepal.Length             2
#> 308   89 Petal.Length             0
#> 309   89  Petal.Width             2
#> 310   89 Sepal.Length             3
#> 311   89  Sepal.Width             3
#> 312   90 Petal.Length             1
#> 313   90  Petal.Width             1
#> 314   90 Sepal.Length             0
#> 315   90  Sepal.Width             4
#> 316   91 Petal.Length             0
#> 317   91  Petal.Width             4
#> 318   91 Sepal.Length             3
#> 319   91  Sepal.Width             3
#> 320   92 Petal.Length             0
#> 321   92  Petal.Width             2
#> 322   92  Sepal.Width             4
#> 323   93 Petal.Length             0
#> 324   93  Petal.Width             1
#> 325   93 Sepal.Length             2
#> 326   94 Petal.Length             0
#> 327   94  Petal.Width             3
#> 328   94 Sepal.Length             1
#> 329   94  Sepal.Width             4
#> 330   95 Petal.Length             2
#> 331   95  Petal.Width             0
#> 332   96 Petal.Length             1
#> 333   96  Petal.Width             0
#> 334   96 Sepal.Length             2
#> 335   96  Sepal.Width             3
#> 336   97 Petal.Length             1
#> 337   97  Petal.Width             0
#> 338   97 Sepal.Length             3
#> 339   97  Sepal.Width             2
#> 340   98 Petal.Length             0
#> 341   98  Petal.Width             1
#> 342   98  Sepal.Width             2
#> 343   99 Petal.Length             3
#> 344   99  Petal.Width             1
#> 345   99 Sepal.Length             0
#> 346  100 Petal.Length             1
#> 347  100  Petal.Width             0
#> 348  100 Sepal.Length             1
#> 349  100  Sepal.Width             2

min_depth_distribution(ranger::ranger(Species ~ ., data = iris, num.trees = 100))

#>     tree     variable minimal_depth
#> 1      1 Petal.Length             0
#> 2      1  Petal.Width             2
#> 3      1 Sepal.Length             2
#> 4      2 Petal.Length             2
#> 5      2  Petal.Width             0
#> 6      2 Sepal.Length             4
#> 7      2  Sepal.Width             3
#> 8      3 Petal.Length             1
#> 9      3  Petal.Width             0
#> 10     3 Sepal.Length             2
#> 11     3  Sepal.Width             3
#> 12     4 Petal.Length             2
#> 13     4  Petal.Width             0
#> 14     4 Sepal.Length             1
#> 15     4  Sepal.Width             3
#> 16     5  Petal.Width             0
#> 17     5 Sepal.Length             3
#> 18     6 Petal.Length             1
#> 19     6  Petal.Width             0
#> 20     6 Sepal.Length             2
#> 21     6  Sepal.Width             3
#> 22     7 Petal.Length             0
#> 23     7  Petal.Width             2
#> 24     7 Sepal.Length             1
#> 25     7  Sepal.Width             2
#> 26     8 Petal.Length             1
#> 27     8  Petal.Width             1
#> 28     8 Sepal.Length             0
#> 29     8  Sepal.Width             2
#> 30     9 Petal.Length             0
#> 31     9  Petal.Width             1
#> 32     9 Sepal.Length             3
#> 33     9  Sepal.Width             2
#> 34    10 Petal.Length             2
#> 35    10  Petal.Width             1
#> 36    10 Sepal.Length             0
#> 37    11 Petal.Length             3
#> 38    11  Petal.Width             1
#> 39    11 Sepal.Length             0
#> 40    11  Sepal.Width             3
#> 41    12 Petal.Length             1
#> 42    12  Petal.Width             0
#> 43    12  Sepal.Width             3
#> 44    13 Petal.Length             2
#> 45    13  Petal.Width             3
#> 46    13 Sepal.Length             0
#> 47    13  Sepal.Width             1
#> 48    14 Petal.Length             0
#> 49    14  Petal.Width             3
#> 50    14 Sepal.Length             2
#> 51    14  Sepal.Width             2
#> 52    15 Petal.Length             2
#> 53    15  Petal.Width             0
#> 54    15 Sepal.Length             3
#> 55    15  Sepal.Width             2
#> 56    16 Petal.Length             0
#> 57    16  Petal.Width             3
#> 58    16 Sepal.Length             4
#> 59    16  Sepal.Width             4
#> 60    17 Petal.Length             2
#> 61    17  Petal.Width             1
#> 62    17 Sepal.Length             0
#> 63    17  Sepal.Width             2
#> 64    18 Petal.Length             2
#> 65    18  Petal.Width             0
#> 66    18 Sepal.Length             1
#> 67    18  Sepal.Width             4
#> 68    19 Petal.Length             1
#> 69    19  Petal.Width             0
#> 70    19 Sepal.Length             2
#> 71    19  Sepal.Width             2
#> 72    20 Petal.Length             0
#> 73    20  Petal.Width             1
#> 74    20 Sepal.Length             2
#> 75    20  Sepal.Width             3
#> 76    21 Petal.Length             2
#> 77    21  Petal.Width             1
#> 78    21 Sepal.Length             0
#> 79    21  Sepal.Width             3
#> 80    22 Petal.Length             0
#> 81    22  Petal.Width             2
#> 82    22 Sepal.Length             3
#> 83    23 Petal.Length             1
#> 84    23  Petal.Width             0
#> 85    23  Sepal.Width             4
#> 86    24 Petal.Length             2
#> 87    24  Petal.Width             0
#> 88    24  Sepal.Width             3
#> 89    25 Petal.Length             0
#> 90    25  Petal.Width             2
#> 91    25 Sepal.Length             1
#> 92    26 Petal.Length             1
#> 93    26  Petal.Width             0
#> 94    26 Sepal.Length             2
#> 95    26  Sepal.Width             3
#> 96    27 Petal.Length             1
#> 97    27  Petal.Width             0
#> 98    27 Sepal.Length             2
#> 99    27  Sepal.Width             3
#> 100   28 Petal.Length             1
#> 101   28  Petal.Width             0
#> 102   28 Sepal.Length             4
#> 103   28  Sepal.Width             3
#> 104   29 Petal.Length             0
#> 105   30 Petal.Length             2
#> 106   30  Petal.Width             0
#> 107   31 Petal.Length             0
#> 108   31  Petal.Width             1
#> 109   31 Sepal.Length             4
#> 110   31  Sepal.Width             2
#> 111   32 Petal.Length             3
#> 112   32  Petal.Width             0
#> 113   32 Sepal.Length             1
#> 114   32  Sepal.Width             3
#> 115   33 Petal.Length             1
#> 116   33  Petal.Width             1
#> 117   33 Sepal.Length             0
#> 118   33  Sepal.Width             5
#> 119   34 Petal.Length             1
#> 120   34  Petal.Width             1
#> 121   34 Sepal.Length             0
#> 122   34  Sepal.Width             3
#> 123   35 Petal.Length             0
#> 124   35  Petal.Width             1
#> 125   35  Sepal.Width             3
#> 126   36 Petal.Length             0
#> 127   36  Petal.Width             1
#> 128   36  Sepal.Width             3
#> 129   37 Petal.Length             2
#> 130   37  Petal.Width             0
#> 131   37 Sepal.Length             1
#> 132   37  Sepal.Width             3
#> 133   38 Petal.Length             0
#> 134   38  Petal.Width             1
#> 135   38  Sepal.Width             2
#> 136   39 Petal.Length             0
#> 137   39  Petal.Width             2
#> 138   40 Petal.Length             3
#> 139   40  Petal.Width             0
#> 140   40 Sepal.Length             1
#> 141   41 Petal.Length             2
#> 142   41  Petal.Width             0
#> 143   41 Sepal.Length             1
#> 144   42 Petal.Length             0
#> 145   42  Petal.Width             1
#> 146   42 Sepal.Length             3
#> 147   43 Petal.Length             1
#> 148   43  Petal.Width             0
#> 149   43 Sepal.Length             2
#> 150   44 Petal.Length             0
#> 151   44  Petal.Width             2
#> 152   44  Sepal.Width             4
#> 153   45 Petal.Length             0
#> 154   45  Petal.Width             2
#> 155   45  Sepal.Width             3
#> 156   46 Petal.Length             1
#> 157   46  Petal.Width             2
#> 158   46 Sepal.Length             0
#> 159   46  Sepal.Width             1
#> 160   47 Petal.Length             1
#> 161   47  Petal.Width             0
#> 162   47 Sepal.Length             4
#> 163   47  Sepal.Width             2
#> 164   48 Petal.Length             0
#> 165   48  Petal.Width             2
#> 166   48  Sepal.Width             4
#> 167   49  Petal.Width             0
#> 168   49 Sepal.Length             3
#> 169   49  Sepal.Width             2
#> 170   50 Petal.Length             1
#> 171   50  Petal.Width             3
#> 172   50 Sepal.Length             0
#> 173   50  Sepal.Width             2
#> 174   51 Petal.Length             0
#> 175   51  Petal.Width             1
#> 176   52 Petal.Length             0
#> 177   52  Petal.Width             2
#> 178   52 Sepal.Length             3
#> 179   53 Petal.Length             0
#> 180   53  Petal.Width             1
#> 181   54 Petal.Length             0
#> 182   54  Petal.Width             1
#> 183   54 Sepal.Length             3
#> 184   55 Petal.Length             0
#> 185   55  Petal.Width             2
#> 186   56 Petal.Length             0
#> 187   56  Petal.Width             2
#> 188   56 Sepal.Length             4
#> 189   56  Sepal.Width             3
#> 190   57 Petal.Length             0
#> 191   57  Petal.Width             2
#> 192   58 Petal.Length             0
#> 193   58  Petal.Width             1
#> 194   58 Sepal.Length             2
#> 195   58  Sepal.Width             4
#> 196   59 Petal.Length             2
#> 197   59  Petal.Width             0
#> 198   59  Sepal.Width             3
#> 199   60 Petal.Length             0
#> 200   60  Petal.Width             2
#> 201   60 Sepal.Length             2
#> 202   60  Sepal.Width             5
#> 203   61 Petal.Length             1
#> 204   61  Petal.Width             1
#> 205   61 Sepal.Length             0
#> 206   61  Sepal.Width             2
#> 207   62 Petal.Length             3
#> 208   62  Petal.Width             0
#> 209   62 Sepal.Length             2
#> 210   62  Sepal.Width             2
#> 211   63 Petal.Length             1
#> 212   63  Petal.Width             0
#> 213   63 Sepal.Length             4
#> 214   63  Sepal.Width             3
#> 215   64 Petal.Length             0
#> 216   64  Petal.Width             2
#> 217   64 Sepal.Length             1
#> 218   64  Sepal.Width             2
#> 219   65 Petal.Length             2
#> 220   65  Petal.Width             1
#> 221   65 Sepal.Length             0
#> 222   66 Petal.Length             1
#> 223   66  Petal.Width             0
#> 224   66 Sepal.Length             4
#> 225   66  Sepal.Width             2
#> 226   67 Petal.Length             0
#> 227   67  Petal.Width             2
#> 228   67 Sepal.Length             4
#> 229   68 Petal.Length             2
#> 230   68  Petal.Width             2
#> 231   68 Sepal.Length             0
#> 232   68  Sepal.Width             1
#> 233   69 Petal.Length             1
#> 234   69  Petal.Width             2
#> 235   69 Sepal.Length             0
#> 236   69  Sepal.Width             2
#> 237   70 Petal.Length             0
#> 238   70  Petal.Width             2
#> 239   70 Sepal.Length             3
#> 240   71 Petal.Length             1
#> 241   71  Petal.Width             1
#> 242   71 Sepal.Length             0
#> 243   72 Petal.Length             3
#> 244   72  Petal.Width             0
#> 245   72 Sepal.Length             2
#> 246   73 Petal.Length             1
#> 247   73  Petal.Width             2
#> 248   73 Sepal.Length             0
#> 249   74 Petal.Length             2
#> 250   74  Petal.Width             0
#> 251   74 Sepal.Length             4
#> 252   74  Sepal.Width             3
#> 253   75 Petal.Length             2
#> 254   75  Petal.Width             0
#> 255   75 Sepal.Length             4
#> 256   76 Petal.Length             3
#> 257   76  Petal.Width             0
#> 258   76 Sepal.Length             1
#> 259   76  Sepal.Width             3
#> 260   77 Petal.Length             1
#> 261   77  Petal.Width             1
#> 262   77 Sepal.Length             0
#> 263   78 Petal.Length             2
#> 264   78  Petal.Width             0
#> 265   78  Sepal.Width             2
#> 266   79 Petal.Length             2
#> 267   79  Petal.Width             0
#> 268   79 Sepal.Length             1
#> 269   79  Sepal.Width             3
#> 270   80 Petal.Length             2
#> 271   80  Petal.Width             0
#> 272   80 Sepal.Length             2
#> 273   80  Sepal.Width             3
#> 274   81 Petal.Length             2
#> 275   81  Petal.Width             0
#> 276   81 Sepal.Length             3
#> 277   81  Sepal.Width             2
#> 278   82 Petal.Length             0
#> 279   82  Petal.Width             2
#> 280   82 Sepal.Length             1
#> 281   82  Sepal.Width             3
#> 282   83 Petal.Length             0
#> 283   83  Petal.Width             2
#> 284   83 Sepal.Length             2
#> 285   83  Sepal.Width             3
#> 286   84 Petal.Length             0
#> 287   84  Petal.Width             2
#> 288   84 Sepal.Length             2
#> 289   84  Sepal.Width             4
#> 290   85 Petal.Length             2
#> 291   85  Petal.Width             0
#> 292   86 Petal.Length             2
#> 293   86  Petal.Width             0
#> 294   87 Petal.Length             0
#> 295   87  Petal.Width             4
#> 296   87  Sepal.Width             2
#> 297   88  Petal.Width             0
#> 298   88 Sepal.Length             2
#> 299   88  Sepal.Width             3
#> 300   89 Petal.Length             2
#> 301   89  Petal.Width             0
#> 302   89 Sepal.Length             3
#> 303   89  Sepal.Width             3
#> 304   90 Petal.Length             1
#> 305   90  Petal.Width             0
#> 306   90 Sepal.Length             3
#> 307   90  Sepal.Width             3
#> 308   91 Petal.Length             0
#> 309   91  Petal.Width             3
#> 310   91 Sepal.Length             2
#> 311   92 Petal.Length             0
#> 312   92  Petal.Width             1
#> 313   92  Sepal.Width             2
#> 314   93 Petal.Length             0
#> 315   93  Petal.Width             2
#> 316   94 Petal.Length             1
#> 317   94  Petal.Width             0
#> 318   94 Sepal.Length             3
#> 319   95 Petal.Length             1
#> 320   95  Petal.Width             4
#> 321   95 Sepal.Length             0
#> 322   95  Sepal.Width             2
#> 323   96 Petal.Length             1
#> 324   96 Sepal.Length             0
#> 325   96  Sepal.Width             2
#> 326   97 Petal.Length             3
#> 327   97  Petal.Width             0
#> 328   97 Sepal.Length             2
#> 329   98 Petal.Length             1
#> 330   98  Petal.Width             0
#> 331   98  Sepal.Width             3
#> 332   99 Petal.Length             1
#> 333   99  Petal.Width             0
#> 334   99  Sepal.Width             3
#> 335  100 Petal.Length             1
#> 336  100  Petal.Width             3
#> 337  100 Sepal.Length             0
#> 338  100  Sepal.Width             4