Explain a random forest

Explains a random forest in a html document using plots created by randomForestExplainer

explain_forest(forest, interactions = FALSE, data = NULL,
  vars = NULL, no_of_pred_plots = 3, pred_grid = 100,
  measures = NULL)

Arguments

forest
A randomForest object created with the option localImp = TRUE

interactions
Logical value: should variable interactions be considered (this may be time-consuming)

data
The data frame on which forest was trained - necessary if interactions = TRUE

vars
A character vector with variables with respect to which interactions will be considered if NULL then they will be selected using the important_variables() function

no_of_pred_plots
The number of most frequent interactions of numeric variables to plot predictions for

pred_grid
The number of points on the grid of plot_predict_interaction (decrease in case memory problems)

measures
A character vector specifying the importance measures to be used for plotting ggpairs

Value

A html document in your working directory

Examples

forest <- randomForest::randomForest(Species ~ ., data = iris, localImp = TRUE)
explain_forest(forest, interactions = TRUE)

#> Error in if (grepl(.shell_chars_regex, basename(input))) {    input_no_shell_chars <- intermediates_loc(file_name_without_shell_chars(basename(input)))    if (file.exists(input_no_shell_chars)) {        stop("The name of the input file cannot contain the special shell ",             "characters: ", .shell_chars_regex, " (attempted to copy to a ",             "version without those characters '", input_no_shell_chars,             "' ", "however that file already exists)", call. = FALSE)    }    file.copy(input, input_no_shell_chars, overwrite = TRUE)    intermediates <- c(intermediates, input_no_shell_chars)    input <- input_no_shell_chars    if (is.null(intermediates_dir)) {        intermediates_dir <- dirname(normalize_path(input_no_shell_chars))        if (same_path(intermediates_dir, dirname(original_input)))             intermediates_dir <- NULL    }}: argument is of length zero