Calculate minimal depth distribution of a random forest

Get minimal depth values for all trees in a random forest

min_depth_distribution(forest)

Arguments

forest

A randomForest or ranger object

Value

A data frame with the value of minimal depth for every variable in every tree

Examples

min_depth_distribution(randomForest::randomForest(Species ~ ., data = iris, ntree = 100))
#>     tree     variable minimal_depth
#> 1      1 Petal.Length             3
#> 2      1  Petal.Width             0
#> 3      1 Sepal.Length             4
#> 4      1  Sepal.Width             2
#> 5      2 Petal.Length             2
#> 6      2  Petal.Width             1
#> 7      2 Sepal.Length             0
#> 8      2  Sepal.Width             2
#> 9      3 Petal.Length             0
#> 10     3  Petal.Width             1
#> 11     3 Sepal.Length             2
#> 12     4 Petal.Length             1
#> 13     4  Petal.Width             0
#> 14     4 Sepal.Length             2
#> 15     4  Sepal.Width             3
#> 16     5 Petal.Length             2
#> 17     5  Petal.Width             1
#> 18     5 Sepal.Length             0
#> 19     6 Petal.Length             2
#> 20     6  Petal.Width             0
#> 21     6 Sepal.Length             2
#> 22     6  Sepal.Width             3
#> 23     7 Petal.Length             1
#> 24     7  Petal.Width             0
#> 25     7 Sepal.Length             2
#> 26     7  Sepal.Width             4
#> 27     8 Petal.Length             0
#> 28     8  Petal.Width             4
#> 29     8 Sepal.Length             1
#> 30     8  Sepal.Width             3
#> 31     9 Petal.Length             3
#> 32     9  Petal.Width             0
#> 33     9 Sepal.Length             3
#> 34     9  Sepal.Width             2
#> 35    10 Petal.Length             1
#> 36    10  Petal.Width             1
#> 37    10 Sepal.Length             0
#> 38    10  Sepal.Width             4
#> 39    11 Petal.Length             2
#> 40    11  Petal.Width             0
#> 41    11 Sepal.Length             1
#> 42    11  Sepal.Width             4
#> 43    12 Petal.Length             1
#> 44    12  Petal.Width             0
#> 45    12 Sepal.Length             4
#> 46    13 Petal.Length             0
#> 47    13  Petal.Width             3
#> 48    13 Sepal.Length             1
#> 49    13  Sepal.Width             4
#> 50    14 Petal.Length             2
#> 51    14  Petal.Width             0
#> 52    14 Sepal.Length             3
#> 53    14  Sepal.Width             4
#> 54    15 Petal.Length             1
#> 55    15  Petal.Width             0
#> 56    15 Sepal.Length             2
#> 57    15  Sepal.Width             3
#> 58    16 Petal.Length             3
#> 59    16  Petal.Width             0
#> 60    16  Sepal.Width             2
#> 61    17 Petal.Length             2
#> 62    17  Petal.Width             0
#> 63    17 Sepal.Length             3
#> 64    17  Sepal.Width             3
#> 65    18 Petal.Length             1
#> 66    18  Petal.Width             0
#> 67    18 Sepal.Length             3
#> 68    18  Sepal.Width             3
#> 69    19 Petal.Length             0
#> 70    19  Petal.Width             1
#> 71    19 Sepal.Length             3
#> 72    19  Sepal.Width             3
#> 73    20 Petal.Length             1
#> 74    20  Petal.Width             0
#> 75    20 Sepal.Length             2
#> 76    21 Petal.Length             0
#> 77    21  Petal.Width             1
#> 78    21 Sepal.Length             3
#> 79    21  Sepal.Width             3
#> 80    22 Petal.Length             1
#> 81    22  Petal.Width             0
#> 82    23 Petal.Length             2
#> 83    23  Petal.Width             0
#> 84    23 Sepal.Length             3
#> 85    23  Sepal.Width             4
#> 86    24 Petal.Length             1
#> 87    24  Petal.Width             1
#> 88    24 Sepal.Length             0
#> 89    25 Petal.Length             1
#> 90    25  Petal.Width             2
#> 91    25 Sepal.Length             0
#> 92    25  Sepal.Width             1
#> 93    26 Petal.Length             1
#> 94    26  Petal.Width             0
#> 95    26 Sepal.Length             5
#> 96    26  Sepal.Width             3
#> 97    27 Petal.Length             2
#> 98    27  Petal.Width             0
#> 99    27  Sepal.Width             3
#> 100   28 Petal.Length             0
#> 101   28  Petal.Width             2
#> 102   28 Sepal.Length             2
#> 103   29 Petal.Length             0
#> 104   29  Petal.Width             1
#> 105   29 Sepal.Length             4
#> 106   30 Petal.Length             1
#> 107   30  Petal.Width             0
#> 108   30 Sepal.Length             3
#> 109   31 Petal.Length             2
#> 110   31  Petal.Width             0
#> 111   31 Sepal.Length             4
#> 112   31  Sepal.Width             3
#> 113   32 Petal.Length             0
#> 114   32  Petal.Width             3
#> 115   32 Sepal.Length             4
#> 116   32  Sepal.Width             3
#> 117   33 Petal.Length             0
#> 118   33  Petal.Width             2
#> 119   34 Petal.Length             0
#> 120   34  Petal.Width             1
#> 121   34 Sepal.Length             2
#> 122   35 Petal.Length             0
#> 123   35  Petal.Width             2
#> 124   35 Sepal.Length             1
#> 125   35  Sepal.Width             3
#> 126   36 Petal.Length             1
#> 127   36  Petal.Width             0
#> 128   36 Sepal.Length             4
#> 129   37 Petal.Length             0
#> 130   37  Petal.Width             1
#> 131   37 Sepal.Length             2
#> 132   37  Sepal.Width             3
#> 133   38 Petal.Length             1
#> 134   38  Petal.Width             2
#> 135   38 Sepal.Length             0
#> 136   38  Sepal.Width             3
#> 137   39 Petal.Length             0
#> 138   39  Petal.Width             1
#> 139   39 Sepal.Length             3
#> 140   39  Sepal.Width             3
#> 141   40 Petal.Length             3
#> 142   40  Petal.Width             0
#> 143   40 Sepal.Length             2
#> 144   40  Sepal.Width             5
#> 145   41 Petal.Length             2
#> 146   41  Petal.Width             0
#> 147   41 Sepal.Length             1
#> 148   41  Sepal.Width             3
#> 149   42 Petal.Length             0
#> 150   42  Petal.Width             4
#> 151   42 Sepal.Length             3
#> 152   42  Sepal.Width             3
#> 153   43 Petal.Length             0
#> 154   43  Petal.Width             1
#> 155   43 Sepal.Length             2
#> 156   44 Petal.Length             0
#> 157   44  Petal.Width             2
#> 158   45 Petal.Length             0
#> 159   45  Petal.Width             1
#> 160   45 Sepal.Length             2
#> 161   45  Sepal.Width             3
#> 162   46 Petal.Length             1
#> 163   46  Petal.Width             0
#> 164   46 Sepal.Length             2
#> 165   46  Sepal.Width             3
#> 166   47 Petal.Length             0
#> 167   47  Petal.Width             2
#> 168   47 Sepal.Length             2
#> 169   47  Sepal.Width             3
#> 170   48 Petal.Length             0
#> 171   48  Petal.Width             1
#> 172   48 Sepal.Length             1
#> 173   48  Sepal.Width             4
#> 174   49 Petal.Length             2
#> 175   49  Petal.Width             0
#> 176   49 Sepal.Length             1
#> 177   50 Petal.Length             2
#> 178   50  Petal.Width             0
#> 179   50 Sepal.Length             3
#> 180   50  Sepal.Width             3
#> 181   51 Petal.Length             2
#> 182   51  Petal.Width             0
#> 183   51 Sepal.Length             4
#> 184   51  Sepal.Width             3
#> 185   52 Petal.Length             1
#> 186   52  Petal.Width             0
#> 187   52 Sepal.Length             3
#> 188   52  Sepal.Width             4
#> 189   53 Petal.Length             0
#> 190   53  Petal.Width             5
#> 191   53 Sepal.Length             3
#> 192   53  Sepal.Width             2
#> 193   54 Petal.Length             0
#> 194   54  Petal.Width             2
#> 195   54 Sepal.Length             1
#> 196   54  Sepal.Width             3
#> 197   55 Petal.Length             1
#> 198   55  Petal.Width             3
#> 199   55 Sepal.Length             0
#> 200   55  Sepal.Width             2
#> 201   56 Petal.Length             2
#> 202   56  Petal.Width             0
#> 203   56 Sepal.Length             2
#> 204   56  Sepal.Width             5
#> 205   57 Petal.Length             0
#> 206   57  Petal.Width             2
#> 207   57 Sepal.Length             3
#> 208   57  Sepal.Width             4
#> 209   58 Petal.Length             0
#> 210   58  Petal.Width             2
#> 211   58  Sepal.Width             4
#> 212   59 Petal.Length             5
#> 213   59  Petal.Width             1
#> 214   59 Sepal.Length             0
#> 215   59  Sepal.Width             4
#> 216   60 Petal.Length             1
#> 217   60  Petal.Width             0
#> 218   60 Sepal.Length             2
#> 219   60  Sepal.Width             2
#> 220   61 Petal.Length             2
#> 221   61  Petal.Width             0
#> 222   61 Sepal.Length             3
#> 223   62 Petal.Length             2
#> 224   62  Petal.Width             0
#> 225   62 Sepal.Length             1
#> 226   62  Sepal.Width             3
#> 227   63 Petal.Length             2
#> 228   63  Petal.Width             1
#> 229   63 Sepal.Length             0
#> 230   63  Sepal.Width             3
#> 231   64 Petal.Length             1
#> 232   64  Petal.Width             1
#> 233   64 Sepal.Length             0
#> 234   65  Petal.Width             0
#> 235   65 Sepal.Length             2
#> 236   65  Sepal.Width             3
#> 237   66 Petal.Length             2
#> 238   66  Petal.Width             1
#> 239   66 Sepal.Length             0
#> 240   66  Sepal.Width             2
#> 241   67 Petal.Length             0
#> 242   67  Petal.Width             1
#> 243   67 Sepal.Length             2
#> 244   67  Sepal.Width             3
#> 245   68 Petal.Length             0
#> 246   68  Petal.Width             1
#> 247   68  Sepal.Width             3
#> 248   69 Petal.Length             1
#> 249   69  Petal.Width             0
#> 250   69 Sepal.Length             4
#> 251   69  Sepal.Width             3
#> 252   70 Petal.Length             3
#> 253   70  Petal.Width             1
#> 254   70 Sepal.Length             0
#> 255   70  Sepal.Width             3
#> 256   71 Petal.Length             0
#> 257   71  Petal.Width             2
#> 258   71 Sepal.Length             3
#> 259   71  Sepal.Width             4
#> 260   72 Petal.Length             0
#> 261   72  Petal.Width             2
#> 262   72 Sepal.Length             1
#> 263   73 Petal.Length             0
#> 264   73  Petal.Width             1
#> 265   73 Sepal.Length             3
#> 266   73  Sepal.Width             5
#> 267   74 Petal.Length             0
#> 268   74  Petal.Width             2
#> 269   74 Sepal.Length             2
#> 270   74  Sepal.Width             3
#> 271   75 Petal.Length             0
#> 272   75  Petal.Width             1
#> 273   75 Sepal.Length             2
#> 274   75  Sepal.Width             3
#> 275   76 Petal.Length             1
#> 276   76  Petal.Width             0
#> 277   76 Sepal.Length             2
#> 278   77 Petal.Length             1
#> 279   77  Petal.Width             1
#> 280   77 Sepal.Length             0
#> 281   77  Sepal.Width             2
#> 282   78 Petal.Length             2
#> 283   78  Petal.Width             0
#> 284   78 Sepal.Length             1
#> 285   79 Petal.Length             0
#> 286   79  Petal.Width             1
#> 287   80 Petal.Length             0
#> 288   80  Petal.Width             2
#> 289   80 Sepal.Length             1
#> 290   80  Sepal.Width             3
#> 291   81 Petal.Length             1
#> 292   81  Petal.Width             0
#> 293   81 Sepal.Length             2
#> 294   81  Sepal.Width             4
#> 295   82 Petal.Length             2
#> 296   82  Petal.Width             0
#> 297   82 Sepal.Length             1
#> 298   82  Sepal.Width             3
#> 299   83 Petal.Length             1
#> 300   83  Petal.Width             0
#> 301   83 Sepal.Length             2
#> 302   84 Petal.Length             1
#> 303   84  Petal.Width             0
#> 304   84 Sepal.Length             2
#> 305   84  Sepal.Width             3
#> 306   85 Petal.Length             0
#> 307   85  Petal.Width             1
#> 308   85 Sepal.Length             3
#> 309   85  Sepal.Width             2
#> 310   86 Petal.Length             0
#> 311   86  Petal.Width             4
#> 312   86  Sepal.Width             3
#> 313   87 Petal.Length             0
#> 314   87  Petal.Width             1
#> 315   87 Sepal.Length             3
#> 316   87  Sepal.Width             4
#> 317   88 Petal.Length             0
#> 318   88  Petal.Width             4
#> 319   88 Sepal.Length             2
#> 320   88  Sepal.Width             3
#> 321   89 Petal.Length             0
#> 322   89  Petal.Width             2
#> 323   89 Sepal.Length             4
#> 324   90 Petal.Length             1
#> 325   90  Petal.Width             2
#> 326   90 Sepal.Length             0
#> 327   90  Sepal.Width             4
#> 328   91 Petal.Length             1
#> 329   91  Petal.Width             0
#> 330   91 Sepal.Length             3
#> 331   91  Sepal.Width             3
#> 332   92 Petal.Length             0
#> 333   92 Sepal.Length             3
#> 334   92  Sepal.Width             3
#> 335   93 Petal.Length             0
#> 336   93  Petal.Width             1
#> 337   93  Sepal.Width             3
#> 338   94 Petal.Length             2
#> 339   94  Petal.Width             0
#> 340   94 Sepal.Length             1
#> 341   95 Petal.Length             0
#> 342   95  Petal.Width             2
#> 343   95 Sepal.Length             2
#> 344   95  Sepal.Width             3
#> 345   96 Petal.Length             2
#> 346   96  Petal.Width             0
#> 347   97 Petal.Length             2
#> 348   97  Petal.Width             0
#> 349   97 Sepal.Length             2
#> 350   97  Sepal.Width             3
#> 351   98 Petal.Length             3
#> 352   98  Petal.Width             1
#> 353   98 Sepal.Length             0
#> 354   98  Sepal.Width             2
#> 355   99 Petal.Length             0
#> 356   99  Petal.Width             3
#> 357   99 Sepal.Length             1
#> 358   99  Sepal.Width             3
#> 359  100 Petal.Length             1
#> 360  100  Petal.Width             0
#> 361  100  Sepal.Width             2
min_depth_distribution(ranger::ranger(Species ~ ., data = iris, num.trees = 100))
#>     tree     variable minimal_depth
#> 1      1 Petal.Length             3
#> 2      1  Petal.Width             1
#> 3      1 Sepal.Length             0
#> 4      1  Sepal.Width             1
#> 5      2 Petal.Length             0
#> 6      2  Petal.Width             4
#> 7      2 Sepal.Length             1
#> 8      2  Sepal.Width             4
#> 9      3 Petal.Length             0
#> 10     3  Petal.Width             1
#> 11     3 Sepal.Length             2
#> 12     4 Petal.Length             1
#> 13     4  Petal.Width             0
#> 14     4 Sepal.Length             1
#> 15     4  Sepal.Width             2
#> 16     5 Petal.Length             0
#> 17     5  Petal.Width             1
#> 18     5 Sepal.Length             3
#> 19     5  Sepal.Width             3
#> 20     6 Petal.Length             2
#> 21     6  Petal.Width             0
#> 22     6 Sepal.Length             1
#> 23     7 Petal.Length             0
#> 24     7  Petal.Width             2
#> 25     7 Sepal.Length             2
#> 26     7  Sepal.Width             5
#> 27     8 Petal.Length             1
#> 28     8  Petal.Width             0
#> 29     8 Sepal.Length             2
#> 30     9 Petal.Length             0
#> 31     9  Petal.Width             1
#> 32     9 Sepal.Length             3
#> 33     9  Sepal.Width             3
#> 34    10 Petal.Length             1
#> 35    10  Petal.Width             0
#> 36    11 Petal.Length             0
#> 37    11  Petal.Width             1
#> 38    11 Sepal.Length             3
#> 39    11  Sepal.Width             5
#> 40    12 Petal.Length             0
#> 41    12  Petal.Width             2
#> 42    12  Sepal.Width             4
#> 43    13 Petal.Length             3
#> 44    13  Petal.Width             0
#> 45    13 Sepal.Length             1
#> 46    13  Sepal.Width             4
#> 47    14 Petal.Length             4
#> 48    14  Petal.Width             0
#> 49    14 Sepal.Length             2
#> 50    14  Sepal.Width             3
#> 51    15 Petal.Length             0
#> 52    15  Petal.Width             1
#> 53    15 Sepal.Length             2
#> 54    15  Sepal.Width             3
#> 55    16 Petal.Length             1
#> 56    16  Petal.Width             0
#> 57    16 Sepal.Length             3
#> 58    16  Sepal.Width             5
#> 59    17 Petal.Length             0
#> 60    17  Petal.Width             1
#> 61    17 Sepal.Length             3
#> 62    17  Sepal.Width             3
#> 63    18 Petal.Length             2
#> 64    18  Petal.Width             0
#> 65    18  Sepal.Width             3
#> 66    19 Petal.Length             1
#> 67    19  Petal.Width             0
#> 68    20 Petal.Length             2
#> 69    20  Petal.Width             0
#> 70    20  Sepal.Width             1
#> 71    21 Petal.Length             0
#> 72    21  Petal.Width             1
#> 73    21  Sepal.Width             3
#> 74    22 Petal.Length             2
#> 75    22  Petal.Width             0
#> 76    22 Sepal.Length             1
#> 77    22  Sepal.Width             4
#> 78    23 Petal.Length             0
#> 79    23  Petal.Width             2
#> 80    23 Sepal.Length             1
#> 81    24 Petal.Length             0
#> 82    24  Petal.Width             1
#> 83    24 Sepal.Length             4
#> 84    24  Sepal.Width             3
#> 85    25 Petal.Length             2
#> 86    25  Petal.Width             1
#> 87    25 Sepal.Length             0
#> 88    25  Sepal.Width             1
#> 89    26 Petal.Length             1
#> 90    26  Petal.Width             0
#> 91    26 Sepal.Length             2
#> 92    26  Sepal.Width             4
#> 93    27 Petal.Length             1
#> 94    27  Petal.Width             0
#> 95    27 Sepal.Length             2
#> 96    28 Petal.Length             0
#> 97    28  Petal.Width             2
#> 98    28 Sepal.Length             4
#> 99    28  Sepal.Width             3
#> 100   29 Petal.Length             1
#> 101   29  Petal.Width             0
#> 102   29 Sepal.Length             3
#> 103   29  Sepal.Width             3
#> 104   30 Petal.Length             0
#> 105   30  Petal.Width             3
#> 106   30 Sepal.Length             4
#> 107   31 Petal.Length             0
#> 108   31  Petal.Width             1
#> 109   31  Sepal.Width             3
#> 110   32 Petal.Length             0
#> 111   32  Petal.Width             1
#> 112   32  Sepal.Width             3
#> 113   33 Petal.Length             2
#> 114   33  Petal.Width             0
#> 115   34 Petal.Length             0
#> 116   34  Petal.Width             1
#> 117   34  Sepal.Width             2
#> 118   35 Petal.Length             1
#> 119   35  Petal.Width             0
#> 120   35 Sepal.Length             2
#> 121   35  Sepal.Width             3
#> 122   36 Petal.Length             2
#> 123   36  Petal.Width             0
#> 124   36 Sepal.Length             3
#> 125   36  Sepal.Width             4
#> 126   37 Petal.Length             0
#> 127   37  Petal.Width             1
#> 128   37 Sepal.Length             3
#> 129   38 Petal.Length             0
#> 130   38  Petal.Width             1
#> 131   38 Sepal.Length             3
#> 132   39 Petal.Length             0
#> 133   39  Petal.Width             2
#> 134   39 Sepal.Length             4
#> 135   39  Sepal.Width             5
#> 136   40 Petal.Length             0
#> 137   40  Petal.Width             1
#> 138   40 Sepal.Length             3
#> 139   40  Sepal.Width             2
#> 140   41 Petal.Length             2
#> 141   41  Petal.Width             0
#> 142   41 Sepal.Length             1
#> 143   41  Sepal.Width             4
#> 144   42 Petal.Length             1
#> 145   42  Petal.Width             0
#> 146   42 Sepal.Length             3
#> 147   42  Sepal.Width             3
#> 148   43 Petal.Length             0
#> 149   43  Petal.Width             3
#> 150   43 Sepal.Length             2
#> 151   43  Sepal.Width             3
#> 152   44 Petal.Length             1
#> 153   44  Petal.Width             0
#> 154   45 Petal.Length             3
#> 155   45  Petal.Width             0
#> 156   45 Sepal.Length             3
#> 157   45  Sepal.Width             2
#> 158   46 Petal.Length             1
#> 159   46  Petal.Width             0
#> 160   46 Sepal.Length             3
#> 161   47 Petal.Length             1
#> 162   47  Petal.Width             2
#> 163   47 Sepal.Length             0
#> 164   47  Sepal.Width             3
#> 165   48 Petal.Length             2
#> 166   48  Petal.Width             1
#> 167   48 Sepal.Length             0
#> 168   48  Sepal.Width             3
#> 169   49 Petal.Length             0
#> 170   49  Petal.Width             1
#> 171   49 Sepal.Length             3
#> 172   49  Sepal.Width             4
#> 173   50 Petal.Length             0
#> 174   50  Petal.Width             2
#> 175   50 Sepal.Length             1
#> 176   50  Sepal.Width             6
#> 177   51 Petal.Length             0
#> 178   51  Petal.Width             2
#> 179   51 Sepal.Length             3
#> 180   51  Sepal.Width             4
#> 181   52 Petal.Length             2
#> 182   52  Petal.Width             0
#> 183   52 Sepal.Length             3
#> 184   53 Petal.Length             3
#> 185   53  Petal.Width             0
#> 186   54 Petal.Length             0
#> 187   54  Petal.Width             1
#> 188   54 Sepal.Length             3
#> 189   54  Sepal.Width             4
#> 190   55 Petal.Length             2
#> 191   55  Petal.Width             0
#> 192   55 Sepal.Length             1
#> 193   56 Petal.Length             0
#> 194   56  Petal.Width             2
#> 195   56 Sepal.Length             1
#> 196   57 Petal.Length             0
#> 197   57  Petal.Width             2
#> 198   57  Sepal.Width             4
#> 199   58 Petal.Length             0
#> 200   58  Petal.Width             3
#> 201   58 Sepal.Length             2
#> 202   59 Petal.Length             1
#> 203   59  Petal.Width             0
#> 204   59 Sepal.Length             4
#> 205   59  Sepal.Width             4
#> 206   60 Petal.Length             2
#> 207   60  Petal.Width             0
#> 208   60  Sepal.Width             3
#> 209   61 Petal.Length             1
#> 210   61  Petal.Width             0
#> 211   61 Sepal.Length             3
#> 212   61  Sepal.Width             2
#> 213   62 Petal.Length             1
#> 214   62  Petal.Width             0
#> 215   63 Petal.Length             2
#> 216   63  Petal.Width             0
#> 217   63 Sepal.Length             1
#> 218   64 Petal.Length             1
#> 219   64  Petal.Width             1
#> 220   64 Sepal.Length             0
#> 221   64  Sepal.Width             4
#> 222   65 Petal.Length             0
#> 223   65  Petal.Width             2
#> 224   65 Sepal.Length             3
#> 225   65  Sepal.Width             4
#> 226   66 Petal.Length             0
#> 227   66  Petal.Width             1
#> 228   67 Petal.Length             1
#> 229   67  Petal.Width             0
#> 230   67 Sepal.Length             3
#> 231   68 Petal.Length             1
#> 232   68  Petal.Width             2
#> 233   68 Sepal.Length             0
#> 234   68  Sepal.Width             1
#> 235   69 Petal.Length             2
#> 236   69  Petal.Width             0
#> 237   69 Sepal.Length             1
#> 238   69  Sepal.Width             3
#> 239   70 Petal.Length             1
#> 240   70  Petal.Width             0
#> 241   70 Sepal.Length             2
#> 242   71 Petal.Length             0
#> 243   71  Petal.Width             2
#> 244   71  Sepal.Width             3
#> 245   72 Petal.Length             1
#> 246   72  Petal.Width             0
#> 247   72  Sepal.Width             4
#> 248   73 Petal.Length             0
#> 249   73  Petal.Width             2
#> 250   73 Sepal.Length             4
#> 251   73  Sepal.Width             4
#> 252   74 Petal.Length             2
#> 253   74  Petal.Width             0
#> 254   74  Sepal.Width             3
#> 255   75 Petal.Length             2
#> 256   75  Petal.Width             0
#> 257   75 Sepal.Length             2
#> 258   76 Petal.Length             0
#> 259   76  Petal.Width             1
#> 260   76  Sepal.Width             2
#> 261   77 Petal.Length             0
#> 262   77  Petal.Width             2
#> 263   77 Sepal.Length             1
#> 264   77  Sepal.Width             4
#> 265   78 Petal.Length             1
#> 266   78  Petal.Width             0
#> 267   78  Sepal.Width             5
#> 268   79 Petal.Length             2
#> 269   79  Petal.Width             1
#> 270   79 Sepal.Length             0
#> 271   79  Sepal.Width             3
#> 272   80 Petal.Length             1
#> 273   80  Petal.Width             2
#> 274   80 Sepal.Length             0
#> 275   80  Sepal.Width             1
#> 276   81 Petal.Length             0
#> 277   81  Petal.Width             2
#> 278   81  Sepal.Width             3
#> 279   82 Petal.Length             0
#> 280   82  Petal.Width             1
#> 281   82 Sepal.Length             2
#> 282   83 Petal.Length             3
#> 283   83  Petal.Width             0
#> 284   83 Sepal.Length             2
#> 285   84 Petal.Length             2
#> 286   84  Petal.Width             0
#> 287   84 Sepal.Length             3
#> 288   85 Petal.Length             2
#> 289   85  Petal.Width             0
#> 290   85 Sepal.Length             2
#> 291   85  Sepal.Width             3
#> 292   86 Petal.Length             0
#> 293   86  Petal.Width             1
#> 294   87 Petal.Length             1
#> 295   87  Petal.Width             0
#> 296   87  Sepal.Width             4
#> 297   88 Petal.Length             0
#> 298   88  Petal.Width             1
#> 299   88  Sepal.Width             2
#> 300   89 Petal.Length             0
#> 301   89  Petal.Width             3
#> 302   89 Sepal.Length             1
#> 303   90 Petal.Length             1
#> 304   90  Petal.Width             1
#> 305   90 Sepal.Length             0
#> 306   91 Petal.Length             1
#> 307   91  Petal.Width             0
#> 308   91  Sepal.Width             2
#> 309   92 Petal.Length             0
#> 310   92  Petal.Width             1
#> 311   92 Sepal.Length             2
#> 312   93 Petal.Length             1
#> 313   93  Petal.Width             0
#> 314   93 Sepal.Length             2
#> 315   93  Sepal.Width             5
#> 316   94 Petal.Length             0
#> 317   94  Petal.Width             1
#> 318   94 Sepal.Length             2
#> 319   95 Petal.Length             0
#> 320   95  Petal.Width             2
#> 321   95 Sepal.Length             1
#> 322   95  Sepal.Width             3
#> 323   96 Petal.Length             3
#> 324   96  Petal.Width             0
#> 325   96 Sepal.Length             1
#> 326   97 Petal.Length             1
#> 327   97  Petal.Width             0
#> 328   97 Sepal.Length             3
#> 329   98 Petal.Length             0
#> 330   98  Petal.Width             1
#> 331   98 Sepal.Length             2
#> 332   99 Petal.Length             3
#> 333   99  Petal.Width             0
#> 334   99 Sepal.Length             2
#> 335   99  Sepal.Width             4
#> 336  100 Petal.Length             2
#> 337  100  Petal.Width             0
#> 338  100 Sepal.Length             2
#> 339  100  Sepal.Width             3