Calculate minimal depth distribution of a random forest

Get minimal depth values for all trees in a random forest

min_depth_distribution(forest)

Arguments

forest

A randomForest or ranger object

Value

A data frame with the value of minimal depth for every variable in every tree

Examples

min_depth_distribution(randomForest::randomForest(Species ~ ., data = iris, ntree = 100))
#>     tree     variable minimal_depth
#> 1      1 Petal.Length             2
#> 2      1  Petal.Width             1
#> 3      1 Sepal.Length             0
#> 4      1  Sepal.Width             1
#> 5      2 Petal.Length             0
#> 6      2  Petal.Width             2
#> 7      2 Sepal.Length             2
#> 8      2  Sepal.Width             3
#> 9      3 Petal.Length             2
#> 10     3  Petal.Width             0
#> 11     3  Sepal.Width             3
#> 12     4 Petal.Length             2
#> 13     4  Petal.Width             1
#> 14     4 Sepal.Length             0
#> 15     4  Sepal.Width             3
#> 16     5 Petal.Length             2
#> 17     5  Petal.Width             0
#> 18     5 Sepal.Length             2
#> 19     5  Sepal.Width             3
#> 20     6 Petal.Length             0
#> 21     6  Petal.Width             3
#> 22     6 Sepal.Length             2
#> 23     6  Sepal.Width             2
#> 24     7 Petal.Length             3
#> 25     7  Petal.Width             0
#> 26     7 Sepal.Length             2
#> 27     7  Sepal.Width             2
#> 28     8 Petal.Length             1
#> 29     8  Petal.Width             2
#> 30     8 Sepal.Length             0
#> 31     8  Sepal.Width             2
#> 32     9 Petal.Length             0
#> 33     9  Petal.Width             1
#> 34     9 Sepal.Length             2
#> 35     9  Sepal.Width             3
#> 36    10 Petal.Length             0
#> 37    10  Petal.Width             1
#> 38    10 Sepal.Length             2
#> 39    10  Sepal.Width             3
#> 40    11 Petal.Length             1
#> 41    11  Petal.Width             0
#> 42    11 Sepal.Length             2
#> 43    11  Sepal.Width             5
#> 44    12 Petal.Length             0
#> 45    12  Petal.Width             1
#> 46    13 Petal.Length             0
#> 47    13  Petal.Width             1
#> 48    13 Sepal.Length             3
#> 49    13  Sepal.Width             3
#> 50    14 Petal.Length             0
#> 51    14  Petal.Width             3
#> 52    14 Sepal.Length             1
#> 53    15 Petal.Length             0
#> 54    15  Petal.Width             3
#> 55    15 Sepal.Length             1
#> 56    15  Sepal.Width             4
#> 57    16 Petal.Length             1
#> 58    16  Petal.Width             1
#> 59    16 Sepal.Length             0
#> 60    17 Petal.Length             4
#> 61    17  Petal.Width             1
#> 62    17 Sepal.Length             0
#> 63    17  Sepal.Width             3
#> 64    18 Petal.Length             0
#> 65    18  Petal.Width             4
#> 66    18 Sepal.Length             2
#> 67    18  Sepal.Width             2
#> 68    19 Petal.Length             1
#> 69    19  Petal.Width             0
#> 70    19 Sepal.Length             4
#> 71    19  Sepal.Width             5
#> 72    20 Petal.Length             3
#> 73    20  Petal.Width             0
#> 74    20 Sepal.Length             2
#> 75    20  Sepal.Width             3
#> 76    21 Petal.Length             1
#> 77    21  Petal.Width             1
#> 78    21 Sepal.Length             0
#> 79    21  Sepal.Width             3
#> 80    22 Petal.Length             5
#> 81    22  Petal.Width             0
#> 82    22 Sepal.Length             2
#> 83    22  Sepal.Width             4
#> 84    23 Petal.Length             0
#> 85    23  Petal.Width             1
#> 86    23  Sepal.Width             3
#> 87    24 Petal.Length             0
#> 88    24  Petal.Width             1
#> 89    24 Sepal.Length             4
#> 90    24  Sepal.Width             3
#> 91    25 Petal.Length             2
#> 92    25  Petal.Width             0
#> 93    25 Sepal.Length             1
#> 94    26 Petal.Length             0
#> 95    26  Petal.Width             1
#> 96    26 Sepal.Length             3
#> 97    26  Sepal.Width             3
#> 98    27 Petal.Length             1
#> 99    27  Petal.Width             0
#> 100   27 Sepal.Length             2
#> 101   27  Sepal.Width             4
#> 102   28 Petal.Length             2
#> 103   28  Petal.Width             0
#> 104   28 Sepal.Length             1
#> 105   28  Sepal.Width             4
#> 106   29 Petal.Length             1
#> 107   29  Petal.Width             3
#> 108   29 Sepal.Length             0
#> 109   29  Sepal.Width             1
#> 110   30 Petal.Length             1
#> 111   30  Petal.Width             0
#> 112   30 Sepal.Length             3
#> 113   30  Sepal.Width             3
#> 114   31 Petal.Length             2
#> 115   31  Petal.Width             0
#> 116   31 Sepal.Length             3
#> 117   32  Petal.Width             0
#> 118   32 Sepal.Length             2
#> 119   33 Petal.Length             1
#> 120   33  Petal.Width             0
#> 121   33 Sepal.Length             2
#> 122   33  Sepal.Width             3
#> 123   34 Petal.Length             1
#> 124   34  Petal.Width             0
#> 125   34 Sepal.Length             3
#> 126   34  Sepal.Width             6
#> 127   35 Petal.Length             1
#> 128   35  Petal.Width             1
#> 129   35 Sepal.Length             0
#> 130   35  Sepal.Width             4
#> 131   36 Petal.Length             2
#> 132   36  Petal.Width             0
#> 133   37 Petal.Length             2
#> 134   37  Petal.Width             0
#> 135   37 Sepal.Length             1
#> 136   37  Sepal.Width             4
#> 137   38 Petal.Length             0
#> 138   38  Petal.Width             1
#> 139   38 Sepal.Length             3
#> 140   38  Sepal.Width             5
#> 141   39 Petal.Length             0
#> 142   39  Petal.Width             1
#> 143   40 Petal.Length             2
#> 144   40  Petal.Width             0
#> 145   40 Sepal.Length             3
#> 146   40  Sepal.Width             4
#> 147   41 Petal.Length             0
#> 148   41  Petal.Width             1
#> 149   41 Sepal.Length             2
#> 150   41  Sepal.Width             4
#> 151   42 Petal.Length             0
#> 152   42  Petal.Width             2
#> 153   42 Sepal.Length             2
#> 154   43 Petal.Length             0
#> 155   43  Petal.Width             1
#> 156   43 Sepal.Length             3
#> 157   44 Petal.Length             0
#> 158   44  Petal.Width             2
#> 159   44 Sepal.Length             3
#> 160   44  Sepal.Width             3
#> 161   45 Petal.Length             2
#> 162   45  Petal.Width             0
#> 163   45 Sepal.Length             3
#> 164   45  Sepal.Width             3
#> 165   46 Petal.Length             0
#> 166   46  Petal.Width             2
#> 167   46 Sepal.Length             4
#> 168   46  Sepal.Width             2
#> 169   47 Petal.Length             2
#> 170   47  Petal.Width             0
#> 171   47 Sepal.Length             3
#> 172   47  Sepal.Width             2
#> 173   48 Petal.Length             2
#> 174   48  Petal.Width             0
#> 175   48  Sepal.Width             2
#> 176   49 Petal.Length             1
#> 177   49  Petal.Width             0
#> 178   49 Sepal.Length             6
#> 179   49  Sepal.Width             4
#> 180   50 Petal.Length             1
#> 181   50  Petal.Width             0
#> 182   50 Sepal.Length             2
#> 183   50  Sepal.Width             4
#> 184   51 Petal.Length             1
#> 185   51  Petal.Width             0
#> 186   51 Sepal.Length             2
#> 187   52 Petal.Length             0
#> 188   52  Petal.Width             4
#> 189   52 Sepal.Length             2
#> 190   53 Petal.Length             0
#> 191   53  Petal.Width             1
#> 192   53 Sepal.Length             3
#> 193   53  Sepal.Width             4
#> 194   54 Petal.Length             1
#> 195   54  Petal.Width             0
#> 196   54 Sepal.Length             2
#> 197   54  Sepal.Width             3
#> 198   55 Petal.Length             0
#> 199   55  Petal.Width             2
#> 200   55 Sepal.Length             2
#> 201   56 Petal.Length             0
#> 202   56  Petal.Width             1
#> 203   56 Sepal.Length             2
#> 204   57 Petal.Length             0
#> 205   57  Petal.Width             1
#> 206   57 Sepal.Length             4
#> 207   57  Sepal.Width             3
#> 208   58 Petal.Length             2
#> 209   58  Petal.Width             0
#> 210   58 Sepal.Length             3
#> 211   58  Sepal.Width             3
#> 212   59 Petal.Length             1
#> 213   59  Petal.Width             1
#> 214   59 Sepal.Length             0
#> 215   59  Sepal.Width             2
#> 216   60 Petal.Length             0
#> 217   60  Petal.Width             1
#> 218   61 Petal.Length             0
#> 219   61  Petal.Width             1
#> 220   61 Sepal.Length             4
#> 221   61  Sepal.Width             3
#> 222   62 Petal.Length             2
#> 223   62  Petal.Width             0
#> 224   62 Sepal.Length             3
#> 225   62  Sepal.Width             3
#> 226   63 Petal.Length             4
#> 227   63  Petal.Width             0
#> 228   63 Sepal.Length             3
#> 229   64 Petal.Length             0
#> 230   64  Petal.Width             2
#> 231   64 Sepal.Length             3
#> 232   65 Petal.Length             1
#> 233   65  Petal.Width             2
#> 234   65 Sepal.Length             0
#> 235   65  Sepal.Width             1
#> 236   66 Petal.Length             2
#> 237   66  Petal.Width             0
#> 238   66 Sepal.Length             1
#> 239   66  Sepal.Width             4
#> 240   67 Petal.Length             0
#> 241   67  Petal.Width             1
#> 242   67 Sepal.Length             2
#> 243   67  Sepal.Width             3
#> 244   68 Petal.Length             1
#> 245   68  Petal.Width             1
#> 246   68 Sepal.Length             0
#> 247   68  Sepal.Width             2
#> 248   69 Petal.Length             3
#> 249   69  Petal.Width             0
#> 250   69  Sepal.Width             2
#> 251   70 Petal.Length             2
#> 252   70  Petal.Width             0
#> 253   70 Sepal.Length             2
#> 254   71 Petal.Length             2
#> 255   71  Petal.Width             1
#> 256   71 Sepal.Length             0
#> 257   71  Sepal.Width             5
#> 258   72 Petal.Length             0
#> 259   72  Petal.Width             1
#> 260   72 Sepal.Length             3
#> 261   72  Sepal.Width             6
#> 262   73 Petal.Length             0
#> 263   73  Petal.Width             3
#> 264   73 Sepal.Length             1
#> 265   73  Sepal.Width             5
#> 266   74 Petal.Length             1
#> 267   74  Petal.Width             0
#> 268   74 Sepal.Length             3
#> 269   74  Sepal.Width             6
#> 270   75 Petal.Length             0
#> 271   75  Petal.Width             2
#> 272   75 Sepal.Length             1
#> 273   75  Sepal.Width             3
#> 274   76 Petal.Length             2
#> 275   76  Petal.Width             0
#> 276   76 Sepal.Length             3
#> 277   77 Petal.Length             1
#> 278   77  Petal.Width             2
#> 279   77 Sepal.Length             0
#> 280   77  Sepal.Width             2
#> 281   78 Petal.Length             1
#> 282   78  Petal.Width             0
#> 283   78  Sepal.Width             3
#> 284   79 Petal.Length             1
#> 285   79  Petal.Width             0
#> 286   79 Sepal.Length             3
#> 287   79  Sepal.Width             2
#> 288   80 Petal.Length             1
#> 289   80  Petal.Width             3
#> 290   80 Sepal.Length             0
#> 291   80  Sepal.Width             2
#> 292   81 Petal.Length             0
#> 293   81  Petal.Width             2
#> 294   81 Sepal.Length             2
#> 295   81  Sepal.Width             3
#> 296   82 Petal.Length             1
#> 297   82  Petal.Width             1
#> 298   82 Sepal.Length             0
#> 299   82  Sepal.Width             4
#> 300   83 Petal.Length             1
#> 301   83  Petal.Width             1
#> 302   83 Sepal.Length             0
#> 303   83  Sepal.Width             4
#> 304   84 Petal.Length             1
#> 305   84  Petal.Width             3
#> 306   84 Sepal.Length             0
#> 307   84  Sepal.Width             2
#> 308   85 Petal.Length             1
#> 309   85  Petal.Width             0
#> 310   85  Sepal.Width             2
#> 311   86 Petal.Length             1
#> 312   86  Petal.Width             0
#> 313   86 Sepal.Length             3
#> 314   86  Sepal.Width             4
#> 315   87 Petal.Length             0
#> 316   87  Petal.Width             1
#> 317   87 Sepal.Length             3
#> 318   87  Sepal.Width             4
#> 319   88 Petal.Length             0
#> 320   88  Petal.Width             1
#> 321   88 Sepal.Length             2
#> 322   88  Sepal.Width             3
#> 323   89 Petal.Length             0
#> 324   89  Petal.Width             1
#> 325   89 Sepal.Length             2
#> 326   90 Petal.Length             3
#> 327   90  Petal.Width             0
#> 328   90 Sepal.Length             1
#> 329   91 Petal.Length             2
#> 330   91  Petal.Width             1
#> 331   91 Sepal.Length             0
#> 332   91  Sepal.Width             3
#> 333   92 Petal.Length             0
#> 334   92  Petal.Width             2
#> 335   92 Sepal.Length             1
#> 336   93 Petal.Length             2
#> 337   93  Petal.Width             0
#> 338   93 Sepal.Length             3
#> 339   93  Sepal.Width             2
#> 340   94 Petal.Length             0
#> 341   94  Petal.Width             3
#> 342   94 Sepal.Length             2
#> 343   94  Sepal.Width             4
#> 344   95 Petal.Length             0
#> 345   95  Petal.Width             1
#> 346   95 Sepal.Length             2
#> 347   95  Sepal.Width             2
#> 348   96 Petal.Length             1
#> 349   96  Petal.Width             0
#> 350   96 Sepal.Length             3
#> 351   96  Sepal.Width             3
#> 352   97 Petal.Length             0
#> 353   97  Petal.Width             1
#> 354   97  Sepal.Width             3
#> 355   98 Petal.Length             3
#> 356   98  Petal.Width             0
#> 357   98 Sepal.Length             1
#> 358   98  Sepal.Width             2
#> 359   99 Petal.Length             0
#> 360   99  Petal.Width             2
#> 361   99 Sepal.Length             1
#> 362   99  Sepal.Width             5
#> 363  100 Petal.Length             0
#> 364  100  Petal.Width             2
#> 365  100 Sepal.Length             2
min_depth_distribution(ranger::ranger(Species ~ ., data = iris, num.trees = 100))
#>     tree     variable minimal_depth
#> 1      1 Petal.Length             1
#> 2      1  Petal.Width             1
#> 3      1 Sepal.Length             0
#> 4      1  Sepal.Width             3
#> 5      2 Petal.Length             0
#> 6      2  Petal.Width             2
#> 7      2 Sepal.Length             1
#> 8      3 Petal.Length             1
#> 9      3  Petal.Width             2
#> 10     3 Sepal.Length             0
#> 11     3  Sepal.Width             1
#> 12     4 Petal.Length             0
#> 13     4  Petal.Width             1
#> 14     4 Sepal.Length             1
#> 15     5 Petal.Length             0
#> 16     5  Petal.Width             2
#> 17     5 Sepal.Length             1
#> 18     5  Sepal.Width             4
#> 19     6 Petal.Length             1
#> 20     6  Petal.Width             0
#> 21     6 Sepal.Length             3
#> 22     7 Petal.Length             1
#> 23     7  Petal.Width             1
#> 24     7 Sepal.Length             0
#> 25     7  Sepal.Width             2
#> 26     8 Petal.Length             1
#> 27     8  Petal.Width             4
#> 28     8 Sepal.Length             0
#> 29     8  Sepal.Width             1
#> 30     9 Petal.Length             0
#> 31     9  Petal.Width             2
#> 32     9 Sepal.Length             1
#> 33    10 Petal.Length             1
#> 34    10  Petal.Width             0
#> 35    10 Sepal.Length             3
#> 36    10  Sepal.Width             2
#> 37    11 Petal.Length             1
#> 38    11  Petal.Width             2
#> 39    11 Sepal.Length             0
#> 40    11  Sepal.Width             2
#> 41    12  Petal.Width             0
#> 42    12  Sepal.Width             2
#> 43    13 Petal.Length             1
#> 44    13  Petal.Width             0
#> 45    14 Petal.Length             0
#> 46    14  Petal.Width             4
#> 47    14 Sepal.Length             2
#> 48    14  Sepal.Width             1
#> 49    15 Petal.Length             1
#> 50    15  Petal.Width             3
#> 51    15 Sepal.Length             0
#> 52    15  Sepal.Width             3
#> 53    16 Petal.Length             1
#> 54    16  Petal.Width             1
#> 55    16 Sepal.Length             0
#> 56    16  Sepal.Width             2
#> 57    17 Petal.Length             0
#> 58    17  Petal.Width             3
#> 59    17 Sepal.Length             2
#> 60    17  Sepal.Width             2
#> 61    18 Petal.Length             1
#> 62    18  Petal.Width             2
#> 63    18 Sepal.Length             0
#> 64    18  Sepal.Width             2
#> 65    19 Petal.Length             0
#> 66    19  Petal.Width             1
#> 67    19 Sepal.Length             3
#> 68    19  Sepal.Width             4
#> 69    20 Petal.Length             2
#> 70    20  Petal.Width             0
#> 71    21 Petal.Length             3
#> 72    21  Petal.Width             0
#> 73    21 Sepal.Length             2
#> 74    21  Sepal.Width             3
#> 75    22 Petal.Length             1
#> 76    22  Petal.Width             2
#> 77    22 Sepal.Length             0
#> 78    22  Sepal.Width             2
#> 79    23 Petal.Length             0
#> 80    23  Petal.Width             1
#> 81    23 Sepal.Length             5
#> 82    23  Sepal.Width             3
#> 83    24 Petal.Length             0
#> 84    24  Petal.Width             3
#> 85    24  Sepal.Width             4
#> 86    25 Petal.Length             0
#> 87    25  Petal.Width             2
#> 88    25 Sepal.Length             2
#> 89    25  Sepal.Width             4
#> 90    26 Petal.Length             1
#> 91    26  Petal.Width             1
#> 92    26 Sepal.Length             0
#> 93    26  Sepal.Width             6
#> 94    27 Petal.Length             0
#> 95    27  Petal.Width             1
#> 96    27 Sepal.Length             2
#> 97    27  Sepal.Width             3
#> 98    28 Petal.Length             2
#> 99    28  Petal.Width             0
#> 100   28  Sepal.Width             3
#> 101   29 Petal.Length             0
#> 102   29  Petal.Width             1
#> 103   29 Sepal.Length             3
#> 104   29  Sepal.Width             2
#> 105   30 Petal.Length             2
#> 106   30  Petal.Width             2
#> 107   30 Sepal.Length             0
#> 108   30  Sepal.Width             1
#> 109   31 Petal.Length             0
#> 110   31  Petal.Width             2
#> 111   31 Sepal.Length             2
#> 112   31  Sepal.Width             3
#> 113   32 Petal.Length             1
#> 114   32  Petal.Width             0
#> 115   32 Sepal.Length             2
#> 116   32  Sepal.Width             3
#> 117   33 Petal.Length             0
#> 118   33  Petal.Width             2
#> 119   33 Sepal.Length             1
#> 120   33  Sepal.Width             3
#> 121   34 Petal.Length             1
#> 122   34  Petal.Width             0
#> 123   34  Sepal.Width             2
#> 124   35 Petal.Length             3
#> 125   35  Petal.Width             0
#> 126   35 Sepal.Length             2
#> 127   36 Petal.Length             2
#> 128   36  Petal.Width             0
#> 129   36 Sepal.Length             1
#> 130   36  Sepal.Width             3
#> 131   37 Petal.Length             0
#> 132   37  Petal.Width             1
#> 133   37  Sepal.Width             3
#> 134   38 Petal.Length             1
#> 135   38  Petal.Width             0
#> 136   38 Sepal.Length             2
#> 137   38  Sepal.Width             4
#> 138   39 Petal.Length             3
#> 139   39  Petal.Width             0
#> 140   39 Sepal.Length             1
#> 141   39  Sepal.Width             4
#> 142   40 Petal.Length             2
#> 143   40  Petal.Width             0
#> 144   40  Sepal.Width             3
#> 145   41 Petal.Length             1
#> 146   41  Petal.Width             0
#> 147   41 Sepal.Length             2
#> 148   41  Sepal.Width             3
#> 149   42 Petal.Length             1
#> 150   42  Petal.Width             0
#> 151   42 Sepal.Length             2
#> 152   42  Sepal.Width             4
#> 153   43 Petal.Length             0
#> 154   43  Petal.Width             2
#> 155   43 Sepal.Length             1
#> 156   44 Petal.Length             1
#> 157   44  Petal.Width             0
#> 158   44 Sepal.Length             2
#> 159   45 Petal.Length             0
#> 160   45  Petal.Width             3
#> 161   45 Sepal.Length             2
#> 162   45  Sepal.Width             3
#> 163   46 Petal.Length             3
#> 164   46  Petal.Width             0
#> 165   46 Sepal.Length             4
#> 166   46  Sepal.Width             4
#> 167   47 Petal.Length             3
#> 168   47  Petal.Width             0
#> 169   47 Sepal.Length             2
#> 170   47  Sepal.Width             3
#> 171   48 Petal.Length             6
#> 172   48  Petal.Width             1
#> 173   48 Sepal.Length             0
#> 174   48  Sepal.Width             2
#> 175   49 Petal.Length             0
#> 176   49  Petal.Width             1
#> 177   49 Sepal.Length             4
#> 178   49  Sepal.Width             4
#> 179   50 Petal.Length             0
#> 180   50 Sepal.Length             2
#> 181   50  Sepal.Width             3
#> 182   51 Petal.Length             0
#> 183   51  Petal.Width             1
#> 184   51  Sepal.Width             3
#> 185   52 Petal.Length             2
#> 186   52  Petal.Width             1
#> 187   52 Sepal.Length             0
#> 188   52  Sepal.Width             1
#> 189   53 Petal.Length             2
#> 190   53  Petal.Width             1
#> 191   53 Sepal.Length             0
#> 192   53  Sepal.Width             1
#> 193   54 Petal.Length             1
#> 194   54  Petal.Width             1
#> 195   54 Sepal.Length             0
#> 196   54  Sepal.Width             4
#> 197   55 Petal.Length             0
#> 198   55  Petal.Width             1
#> 199   55 Sepal.Length             2
#> 200   55  Sepal.Width             3
#> 201   56 Petal.Length             0
#> 202   56  Petal.Width             1
#> 203   56 Sepal.Length             3
#> 204   57 Petal.Length             3
#> 205   57  Petal.Width             0
#> 206   57 Sepal.Length             2
#> 207   57  Sepal.Width             4
#> 208   58 Petal.Length             1
#> 209   58  Petal.Width             0
#> 210   58 Sepal.Length             2
#> 211   59 Petal.Length             0
#> 212   59  Petal.Width             1
#> 213   59  Sepal.Width             3
#> 214   60 Petal.Length             1
#> 215   60  Petal.Width             1
#> 216   60 Sepal.Length             0
#> 217   60  Sepal.Width             2
#> 218   61 Petal.Length             1
#> 219   61  Petal.Width             0
#> 220   61 Sepal.Length             2
#> 221   62 Petal.Length             0
#> 222   62  Sepal.Width             3
#> 223   63 Petal.Length             3
#> 224   63  Petal.Width             0
#> 225   63 Sepal.Length             2
#> 226   63  Sepal.Width             3
#> 227   64 Petal.Length             0
#> 228   64  Petal.Width             2
#> 229   64 Sepal.Length             3
#> 230   65 Petal.Length             0
#> 231   65  Petal.Width             1
#> 232   65 Sepal.Length             4
#> 233   65  Sepal.Width             3
#> 234   66 Petal.Length             0
#> 235   66  Petal.Width             1
#> 236   66 Sepal.Length             2
#> 237   66  Sepal.Width             3
#> 238   67 Petal.Length             0
#> 239   67  Petal.Width             3
#> 240   67 Sepal.Length             3
#> 241   67  Sepal.Width             3
#> 242   68 Petal.Length             0
#> 243   68  Petal.Width             2
#> 244   68 Sepal.Length             2
#> 245   69 Petal.Length             0
#> 246   69  Petal.Width             2
#> 247   69 Sepal.Length             2
#> 248   69  Sepal.Width             5
#> 249   70 Petal.Length             2
#> 250   70  Petal.Width             0
#> 251   70 Sepal.Length             2
#> 252   71 Petal.Length             3
#> 253   71  Petal.Width             0
#> 254   71 Sepal.Length             1
#> 255   71  Sepal.Width             1
#> 256   72 Petal.Length             2
#> 257   72  Petal.Width             1
#> 258   72 Sepal.Length             0
#> 259   72  Sepal.Width             5
#> 260   73 Petal.Length             0
#> 261   73  Petal.Width             1
#> 262   73 Sepal.Length             2
#> 263   73  Sepal.Width             4
#> 264   74  Petal.Width             0
#> 265   74 Sepal.Length             2
#> 266   75 Petal.Length             2
#> 267   75  Petal.Width             1
#> 268   75 Sepal.Length             0
#> 269   75  Sepal.Width             4
#> 270   76 Petal.Length             0
#> 271   76  Petal.Width             2
#> 272   76 Sepal.Length             1
#> 273   76  Sepal.Width             3
#> 274   77 Petal.Length             3
#> 275   77  Petal.Width             0
#> 276   77 Sepal.Length             1
#> 277   77  Sepal.Width             4
#> 278   78 Petal.Length             0
#> 279   78  Petal.Width             2
#> 280   78 Sepal.Length             3
#> 281   78  Sepal.Width             4
#> 282   79 Petal.Length             2
#> 283   79  Petal.Width             0
#> 284   79 Sepal.Length             1
#> 285   80 Petal.Length             0
#> 286   80  Petal.Width             1
#> 287   80 Sepal.Length             4
#> 288   80  Sepal.Width             3
#> 289   81 Petal.Length             1
#> 290   81  Petal.Width             1
#> 291   81 Sepal.Length             0
#> 292   82 Petal.Length             0
#> 293   82  Petal.Width             2
#> 294   82 Sepal.Length             2
#> 295   82  Sepal.Width             3
#> 296   83 Petal.Length             0
#> 297   83  Petal.Width             1
#> 298   83 Sepal.Length             2
#> 299   83  Sepal.Width             4
#> 300   84 Petal.Length             0
#> 301   84  Petal.Width             1
#> 302   84 Sepal.Length             3
#> 303   84  Sepal.Width             2
#> 304   85 Petal.Length             1
#> 305   85  Petal.Width             2
#> 306   85 Sepal.Length             0
#> 307   85  Sepal.Width             2
#> 308   86 Petal.Length             0
#> 309   86  Petal.Width             1
#> 310   86  Sepal.Width             3
#> 311   87 Petal.Length             1
#> 312   87  Petal.Width             0
#> 313   87 Sepal.Length             4
#> 314   87  Sepal.Width             2
#> 315   88 Petal.Length             2
#> 316   88  Petal.Width             0
#> 317   88 Sepal.Length             4
#> 318   88  Sepal.Width             3
#> 319   89 Petal.Length             1
#> 320   89  Petal.Width             0
#> 321   89 Sepal.Length             3
#> 322   89  Sepal.Width             4
#> 323   90 Petal.Length             5
#> 324   90  Petal.Width             0
#> 325   90 Sepal.Length             3
#> 326   90  Sepal.Width             2
#> 327   91 Petal.Length             3
#> 328   91  Petal.Width             1
#> 329   91 Sepal.Length             0
#> 330   91  Sepal.Width             1
#> 331   92 Petal.Length             2
#> 332   92  Petal.Width             0
#> 333   92 Sepal.Length             1
#> 334   92  Sepal.Width             3
#> 335   93 Petal.Length             3
#> 336   93  Petal.Width             1
#> 337   93 Sepal.Length             0
#> 338   94 Petal.Length             2
#> 339   94  Petal.Width             0
#> 340   94 Sepal.Length             3
#> 341   95 Petal.Length             1
#> 342   95  Petal.Width             0
#> 343   95 Sepal.Length             3
#> 344   95  Sepal.Width             4
#> 345   96 Petal.Length             1
#> 346   96  Petal.Width             2
#> 347   96 Sepal.Length             0
#> 348   96  Sepal.Width             2
#> 349   97 Petal.Length             0
#> 350   97  Petal.Width             1
#> 351   98 Petal.Length             0
#> 352   98  Petal.Width             2
#> 353   98 Sepal.Length             1
#> 354   98  Sepal.Width             4
#> 355   99 Petal.Length             0
#> 356   99  Petal.Width             2
#> 357   99 Sepal.Length             3
#> 358   99  Sepal.Width             4
#> 359  100 Petal.Length             0
#> 360  100  Petal.Width             1
#> 361  100 Sepal.Length             2