Calculate minimal depth distribution of a random forest

Get minimal depth values for all trees in a random forest

min_depth_distribution(forest)

Arguments

forest
A randomForest object

Value

A data frame with the value of minimal depth for every variable in every tree

Examples

min_depth_distribution(randomForest::randomForest(Species ~ ., data = iris))

#>      tree     variable minimal_depth
#> 1       1  Petal.Width             0
#> 2       1 Petal.Length             1
#> 3       1  Sepal.Width             3
#> 4       1 Sepal.Length             4
#> 5       2  Petal.Width             0
#> 6       2 Petal.Length             2
#> 7       2 Sepal.Length             3
#> 8       3  Petal.Width             0
#> 9       3 Petal.Length             1
#> 10      3  Sepal.Width             3
#> 11      4  Petal.Width             0
#> 12      4 Petal.Length             1
#> 13      4 Sepal.Length             3
#> 14      5  Petal.Width             0
#> 15      5 Petal.Length             2
#> 16      5  Sepal.Width             3
#> 17      6  Petal.Width             0
#> 18      6 Petal.Length             1
#> 19      7 Petal.Length             0
#> 20      7  Petal.Width             2
#> 21      7 Sepal.Length             3
#> 22      7  Sepal.Width             3
#> 23      8 Sepal.Length             0
#> 24      8 Petal.Length             1
#> 25      8  Sepal.Width             1
#> 26      8  Petal.Width             2
#> 27      9 Petal.Length             0
#> 28      9 Sepal.Length             1
#> 29      9  Sepal.Width             3
#> 30      9  Petal.Width             4
#> 31     10 Petal.Length             0
#> 32     10  Petal.Width             1
#> 33     10 Sepal.Length             3
#> 34     10  Sepal.Width             4
#> 35     11 Sepal.Length             0
#> 36     11 Petal.Length             1
#> 37     11  Sepal.Width             2
#> 38     11  Petal.Width             4
#> 39     12  Petal.Width             0
#> 40     12 Sepal.Length             1
#> 41     12 Petal.Length             2
#> 42     12  Sepal.Width             4
#> 43     13  Petal.Width             0
#> 44     13 Petal.Length             2
#> 45     13 Sepal.Length             4
#> 46     14 Petal.Length             0
#> 47     14  Petal.Width             1
#> 48     14 Sepal.Length             3
#> 49     15 Sepal.Length             0
#> 50     15  Petal.Width             1
#> 51     15 Petal.Length             2
#> 52     16 Petal.Length             0
#> 53     16  Petal.Width             1
#> 54     16 Sepal.Length             2
#> 55     17 Sepal.Length             0
#> 56     17  Petal.Width             1
#> 57     17  Sepal.Width             2
#> 58     17 Petal.Length             3
#> 59     18  Petal.Width             0
#> 60     18 Petal.Length             2
#> 61     18  Sepal.Width             3
#> 62     18 Sepal.Length             4
#> 63     19 Petal.Length             0
#> 64     19  Petal.Width             2
#> 65     19 Sepal.Length             3
#> 66     20  Petal.Width             0
#> 67     20 Petal.Length             1
#> 68     20  Sepal.Width             3
#> 69     20 Sepal.Length             4
#> 70     21 Sepal.Length             0
#> 71     21 Petal.Length             1
#> 72     21  Petal.Width             1
#> 73     21  Sepal.Width             3
#> 74     22 Petal.Length             0
#> 75     22  Petal.Width             2
#> 76     22 Sepal.Length             3
#> 77     22  Sepal.Width             4
#> 78     23 Sepal.Length             0
#> 79     23  Petal.Width             1
#> 80     23 Petal.Length             1
#> 81     24 Sepal.Length             0
#> 82     24 Petal.Length             1
#> 83     24  Petal.Width             1
#> 84     24  Sepal.Width             3
#> 85     25  Petal.Width             0
#> 86     25 Petal.Length             2
#> 87     26  Petal.Width             0
#> 88     26 Sepal.Length             2
#> 89     26 Petal.Length             3
#> 90     26  Sepal.Width             3
#> 91     27 Sepal.Length             0
#> 92     27  Sepal.Width             1
#> 93     27 Petal.Length             1
#> 94     27  Petal.Width             3
#> 95     28 Petal.Length             0
#> 96     28  Petal.Width             1
#> 97     28  Sepal.Width             4
#> 98     28 Sepal.Length             4
#> 99     29  Petal.Width             0
#> 100    29 Petal.Length             2
#> 101    29 Sepal.Length             2
#> 102    29  Sepal.Width             3
#> 103    30  Petal.Width             0
#> 104    30 Petal.Length             2
#> 105    30 Sepal.Length             3
#> 106    31 Petal.Length             0
#> 107    31  Petal.Width             2
#> 108    31 Sepal.Length             3
#> 109    32 Sepal.Length             0
#> 110    32 Petal.Length             1
#> 111    32  Petal.Width             1
#> 112    32  Sepal.Width             5
#> 113    33 Petal.Length             0
#> 114    33  Petal.Width             1
#> 115    33  Sepal.Width             3
#> 116    33 Sepal.Length             5
#> 117    34  Petal.Width             0
#> 118    34 Petal.Length             1
#> 119    34  Sepal.Width             2
#> 120    34 Sepal.Length             3
#> 121    35  Petal.Width             0
#> 122    35 Petal.Length             2
#> 123    35 Sepal.Length             2
#> 124    35  Sepal.Width             3
#> 125    36  Petal.Width             0
#> 126    36 Sepal.Length             2
#> 127    36 Petal.Length             3
#> 128    37  Petal.Width             0
#> 129    37 Sepal.Length             2
#> 130    37 Petal.Length             2
#> 131    37  Sepal.Width             3
#> 132    38  Petal.Width             0
#> 133    38 Sepal.Length             1
#> 134    38 Petal.Length             2
#> 135    38  Sepal.Width             4
#> 136    39 Petal.Length             0
#> 137    39 Sepal.Length             1
#> 138    39  Petal.Width             3
#> 139    39  Sepal.Width             4
#> 140    40 Petal.Length             0
#> 141    40 Sepal.Length             2
#> 142    40  Petal.Width             2
#> 143    40  Sepal.Width             3
#> 144    41 Sepal.Length             0
#> 145    41  Sepal.Width             1
#> 146    41 Petal.Length             1
#> 147    41  Petal.Width             5
#> 148    42 Petal.Length             0
#> 149    42  Petal.Width             1
#> 150    42  Sepal.Width             3
#> 151    43 Petal.Length             0
#> 152    43  Petal.Width             1
#> 153    43 Sepal.Length             2
#> 154    43  Sepal.Width             3
#> 155    44 Petal.Length             0
#> 156    44  Petal.Width             2
#> 157    44  Sepal.Width             3
#> 158    45 Petal.Length             0
#> 159    45  Petal.Width             2
#> 160    45  Sepal.Width             3
#> 161    45 Sepal.Length             4
#> 162    46 Sepal.Length             0
#> 163    46  Petal.Width             1
#> 164    46 Petal.Length             1
#> 165    47  Petal.Width             0
#> 166    47 Petal.Length             2
#> 167    47 Sepal.Length             3
#> 168    48  Petal.Width             0
#> 169    48 Sepal.Length             1
#> 170    48 Petal.Length             2
#> 171    48  Sepal.Width             2
#> 172    49  Petal.Width             0
#> 173    49 Sepal.Length             2
#> 174    49 Petal.Length             5
#> 175    50  Petal.Width             0
#> 176    50  Sepal.Width             3
#> 177    50 Petal.Length             4
#> 178    51  Petal.Width             0
#> 179    51 Petal.Length             1
#> 180    51  Sepal.Width             3
#> 181    52  Petal.Width             0
#> 182    52 Sepal.Length             1
#> 183    52 Petal.Length             2
#> 184    53 Sepal.Length             0
#> 185    53 Petal.Length             1
#> 186    53  Petal.Width             1
#> 187    54  Petal.Width             0
#> 188    54 Petal.Length             2
#> 189    54 Sepal.Length             3
#> 190    55 Sepal.Length             0
#> 191    55 Petal.Length             1
#> 192    55  Petal.Width             1
#> 193    55  Sepal.Width             2
#> 194    56 Petal.Length             0
#> 195    56  Petal.Width             1
#> 196    56  Sepal.Width             4
#> 197    57  Petal.Width             0
#> 198    57 Petal.Length             1
#> 199    57 Sepal.Length             3
#> 200    58 Petal.Length             0
#> 201    58 Sepal.Length             1
#> 202    58  Petal.Width             2
#> 203    58  Sepal.Width             3
#> 204    59 Sepal.Length             0
#> 205    59  Petal.Width             1
#> 206    59 Petal.Length             1
#> 207    59  Sepal.Width             4
#> 208    60 Sepal.Length             0
#> 209    60  Petal.Width             1
#> 210    60 Petal.Length             2
#> 211    60  Sepal.Width             3
#> 212    61  Petal.Width             0
#> 213    61  Sepal.Width             2
#> 214    61 Petal.Length             3
#> 215    61 Sepal.Length             4
#> 216    62  Petal.Width             0
#> 217    62 Sepal.Length             1
#> 218    62  Sepal.Width             2
#> 219    62 Petal.Length             3
#> 220    63 Sepal.Length             0
#> 221    63  Petal.Width             1
#> 222    63 Petal.Length             1
#> 223    63  Sepal.Width             2
#> 224    64 Petal.Length             0
#> 225    64  Petal.Width             1
#> 226    64 Sepal.Length             2
#> 227    64  Sepal.Width             4
#> 228    65 Petal.Length             0
#> 229    65  Petal.Width             1
#> 230    65 Sepal.Length             2
#> 231    65  Sepal.Width             3
#> 232    66  Petal.Width             0
#> 233    66 Sepal.Length             1
#> 234    66 Petal.Length             2
#> 235    66  Sepal.Width             3
#> 236    67  Petal.Width             0
#> 237    67 Petal.Length             1
#> 238    67 Sepal.Length             3
#> 239    67  Sepal.Width             3
#> 240    68 Petal.Length             0
#> 241    68  Petal.Width             1
#> 242    68 Sepal.Length             2
#> 243    68  Sepal.Width             6
#> 244    69 Petal.Length             0
#> 245    69 Sepal.Length             1
#> 246    69  Petal.Width             2
#> 247    69  Sepal.Width             3
#> 248    70  Petal.Width             0
#> 249    70 Petal.Length             2
#> 250    70  Sepal.Width             3
#> 251    70 Sepal.Length             4
#> 252    71  Petal.Width             0
#> 253    71 Petal.Length             2
#> 254    71  Sepal.Width             3
#> 255    72 Petal.Length             0
#> 256    72 Sepal.Length             2
#> 257    72  Petal.Width             2
#> 258    72  Sepal.Width             3
#> 259    73 Petal.Length             0
#> 260    73  Petal.Width             1
#> 261    73 Sepal.Length             4
#> 262    74 Petal.Length             0
#> 263    74 Sepal.Length             1
#> 264    74  Petal.Width             3
#> 265    74  Sepal.Width             4
#> 266    75  Petal.Width             0
#> 267    75 Petal.Length             1
#> 268    75  Sepal.Width             3
#> 269    76 Sepal.Length             0
#> 270    76  Sepal.Width             1
#> 271    76 Petal.Length             1
#> 272    77  Petal.Width             0
#> 273    77 Petal.Length             2
#> 274    77 Sepal.Length             3
#> 275    77  Sepal.Width             3
#> 276    78  Petal.Width             0
#> 277    78 Petal.Length             2
#> 278    79 Sepal.Length             0
#> 279    79  Petal.Width             1
#> 280    79 Petal.Length             1
#> 281    79  Sepal.Width             2
#> 282    80  Petal.Width             0
#> 283    80 Petal.Length             1
#> 284    80 Sepal.Length             4
#> 285    81  Petal.Width             0
#> 286    81 Petal.Length             1
#> 287    81  Sepal.Width             5
#> 288    82 Sepal.Length             0
#> 289    82  Petal.Width             1
#> 290    82 Petal.Length             2
#> 291    82  Sepal.Width             5
#> 292    83  Petal.Width             0
#> 293    83 Petal.Length             2
#> 294    83  Sepal.Width             3
#> 295    84 Petal.Length             0
#> 296    84  Petal.Width             1
#> 297    84 Sepal.Length             2
#> 298    84  Sepal.Width             3
#> 299    85 Petal.Length             0
#> 300    85 Sepal.Length             1
#> 301    85  Petal.Width             2
#> 302    85  Sepal.Width             3
#> 303    86 Petal.Length             0
#> 304    86 Sepal.Length             2
#> 305    86  Petal.Width             2
#> 306    86  Sepal.Width             3
#> 307    87 Petal.Length             0
#> 308    87  Petal.Width             2
#> 309    87 Sepal.Length             3
#> 310    88 Petal.Length             0
#> 311    88  Petal.Width             1
#> 312    88  Sepal.Width             3
#> 313    88 Sepal.Length             4
#> 314    89 Petal.Length             0
#> 315    89 Sepal.Length             1
#> 316    89  Petal.Width             3
#> 317    89  Sepal.Width             5
#> 318    90 Petal.Length             0
#> 319    90 Sepal.Length             1
#> 320    90  Sepal.Width             1
#> 321    90  Petal.Width             2
#> 322    91 Sepal.Length             0
#> 323    91  Sepal.Width             1
#> 324    91 Petal.Length             1
#> 325    91  Petal.Width             2
#> 326    92  Petal.Width             0
#> 327    92 Petal.Length             1
#> 328    92  Sepal.Width             3
#> 329    92 Sepal.Length             3
#> 330    93 Petal.Length             0
#> 331    93  Petal.Width             2
#> 332    93  Sepal.Width             3
#> 333    93 Sepal.Length             4
#>  [ osiągnięto getOption("max.print") -- pominięto 1429 wierszy]