Calculate minimal depth distribution of a random forest

Get minimal depth values for all trees in a random forest

min_depth_distribution(forest)

Arguments

forest

A randomForest or ranger object

Value

A data frame with the value of minimal depth for every variable in every tree

Examples

min_depth_distribution(randomForest::randomForest(Species ~ ., data = iris, ntree = 100))
#>     tree     variable minimal_depth
#> 1      1 Petal.Length             0
#> 2      1  Petal.Width             2
#> 3      1 Sepal.Length             2
#> 4      2 Petal.Length             2
#> 5      2  Petal.Width             1
#> 6      2 Sepal.Length             0
#> 7      2  Sepal.Width             2
#> 8      3 Petal.Length             3
#> 9      3  Petal.Width             1
#> 10     3 Sepal.Length             0
#> 11     3  Sepal.Width             2
#> 12     4 Petal.Length             0
#> 13     4  Petal.Width             2
#> 14     4 Sepal.Length             3
#> 15     5 Petal.Length             0
#> 16     5  Petal.Width             2
#> 17     5 Sepal.Length             2
#> 18     5  Sepal.Width             3
#> 19     6 Petal.Length             0
#> 20     6  Petal.Width             2
#> 21     7 Petal.Length             2
#> 22     7  Petal.Width             0
#> 23     7 Sepal.Length             4
#> 24     7  Sepal.Width             3
#> 25     8 Petal.Length             0
#> 26     8  Petal.Width             4
#> 27     8 Sepal.Length             4
#> 28     8  Sepal.Width             2
#> 29     9 Petal.Length             3
#> 30     9  Petal.Width             0
#> 31     9 Sepal.Length             2
#> 32     9  Sepal.Width             3
#> 33    10 Petal.Length             2
#> 34    10  Petal.Width             0
#> 35    10 Sepal.Length             1
#> 36    10  Sepal.Width             2
#> 37    11 Petal.Length             1
#> 38    11  Petal.Width             1
#> 39    11 Sepal.Length             0
#> 40    11  Sepal.Width             5
#> 41    12 Petal.Length             5
#> 42    12  Petal.Width             0
#> 43    12 Sepal.Length             2
#> 44    12  Sepal.Width             4
#> 45    13 Petal.Length             0
#> 46    13  Petal.Width             2
#> 47    14 Petal.Length             2
#> 48    14  Petal.Width             0
#> 49    14 Sepal.Length             2
#> 50    15 Petal.Length             4
#> 51    15  Petal.Width             0
#> 52    15 Sepal.Length             2
#> 53    16 Petal.Length             0
#> 54    16  Petal.Width             1
#> 55    16 Sepal.Length             2
#> 56    17 Petal.Length             2
#> 57    17  Petal.Width             0
#> 58    17 Sepal.Length             2
#> 59    18 Petal.Length             0
#> 60    18  Petal.Width             2
#> 61    18 Sepal.Length             4
#> 62    18  Sepal.Width             3
#> 63    19 Petal.Length             1
#> 64    19  Petal.Width             0
#> 65    19 Sepal.Length             2
#> 66    19  Sepal.Width             4
#> 67    20 Petal.Length             1
#> 68    20  Petal.Width             0
#> 69    20 Sepal.Length             4
#> 70    20  Sepal.Width             3
#> 71    21 Petal.Length             2
#> 72    21  Petal.Width             0
#> 73    21  Sepal.Width             3
#> 74    22 Petal.Length             3
#> 75    22  Petal.Width             0
#> 76    22 Sepal.Length             2
#> 77    23  Petal.Width             0
#> 78    23  Sepal.Width             2
#> 79    24 Petal.Length             2
#> 80    24  Petal.Width             0
#> 81    24  Sepal.Width             3
#> 82    25 Petal.Length             0
#> 83    25  Petal.Width             1
#> 84    26 Petal.Length             0
#> 85    26  Petal.Width             2
#> 86    26 Sepal.Length             2
#> 87    27 Petal.Length             0
#> 88    27  Petal.Width             2
#> 89    27 Sepal.Length             2
#> 90    28 Petal.Length             0
#> 91    28  Petal.Width             1
#> 92    28 Sepal.Length             2
#> 93    28  Sepal.Width             3
#> 94    29 Petal.Length             2
#> 95    29  Petal.Width             0
#> 96    30 Petal.Length             0
#> 97    30  Petal.Width             3
#> 98    30 Sepal.Length             1
#> 99    30  Sepal.Width             3
#> 100   31 Petal.Length             0
#> 101   31  Petal.Width             1
#> 102   31  Sepal.Width             3
#> 103   32 Petal.Length             3
#> 104   32  Petal.Width             0
#> 105   32 Sepal.Length             2
#> 106   32  Sepal.Width             4
#> 107   33 Petal.Length             0
#> 108   33  Petal.Width             1
#> 109   33 Sepal.Length             2
#> 110   33  Sepal.Width             3
#> 111   34 Petal.Length             2
#> 112   34  Petal.Width             0
#> 113   34 Sepal.Length             1
#> 114   35 Petal.Length             0
#> 115   35  Petal.Width             1
#> 116   35  Sepal.Width             3
#> 117   36 Petal.Length             0
#> 118   36  Petal.Width             1
#> 119   36 Sepal.Length             2
#> 120   36  Sepal.Width             3
#> 121   37 Petal.Length             3
#> 122   37  Petal.Width             0
#> 123   37 Sepal.Length             3
#> 124   37  Sepal.Width             1
#> 125   38 Petal.Length             3
#> 126   38  Petal.Width             0
#> 127   38 Sepal.Length             1
#> 128   38  Sepal.Width             2
#> 129   39 Petal.Length             1
#> 130   39  Petal.Width             1
#> 131   39 Sepal.Length             0
#> 132   40 Petal.Length             1
#> 133   40  Petal.Width             0
#> 134   40 Sepal.Length             2
#> 135   40  Sepal.Width             2
#> 136   41 Petal.Length             2
#> 137   41  Petal.Width             3
#> 138   41 Sepal.Length             0
#> 139   41  Sepal.Width             1
#> 140   42 Petal.Length             1
#> 141   42  Petal.Width             0
#> 142   43 Petal.Length             2
#> 143   43  Petal.Width             0
#> 144   43 Sepal.Length             3
#> 145   43  Sepal.Width             2
#> 146   44 Petal.Length             2
#> 147   44  Petal.Width             0
#> 148   44 Sepal.Length             2
#> 149   44  Sepal.Width             3
#> 150   45 Petal.Length             1
#> 151   45  Petal.Width             0
#> 152   46 Petal.Length             0
#> 153   46  Petal.Width             1
#> 154   46 Sepal.Length             2
#> 155   46  Sepal.Width             3
#> 156   47 Petal.Length             2
#> 157   47  Petal.Width             1
#> 158   47 Sepal.Length             0
#> 159   47  Sepal.Width             1
#> 160   48 Petal.Length             0
#> 161   48  Petal.Width             1
#> 162   48  Sepal.Width             3
#> 163   49 Petal.Length             0
#> 164   49  Petal.Width             1
#> 165   49 Sepal.Length             3
#> 166   49  Sepal.Width             4
#> 167   50 Petal.Length             3
#> 168   50  Petal.Width             0
#> 169   50 Sepal.Length             2
#> 170   50  Sepal.Width             2
#> 171   51 Petal.Length             1
#> 172   51  Petal.Width             0
#> 173   51 Sepal.Length             3
#> 174   51  Sepal.Width             2
#> 175   52 Petal.Length             1
#> 176   52  Petal.Width             1
#> 177   52 Sepal.Length             0
#> 178   52  Sepal.Width             4
#> 179   53 Petal.Length             0
#> 180   53  Petal.Width             2
#> 181   53 Sepal.Length             1
#> 182   54 Petal.Length             0
#> 183   54  Petal.Width             1
#> 184   54 Sepal.Length             2
#> 185   54  Sepal.Width             2
#> 186   55 Petal.Length             0
#> 187   55  Petal.Width             2
#> 188   55 Sepal.Length             1
#> 189   56 Petal.Length             0
#> 190   56  Petal.Width             3
#> 191   56 Sepal.Length             2
#> 192   56  Sepal.Width             5
#> 193   57 Petal.Length             1
#> 194   57  Petal.Width             1
#> 195   57 Sepal.Length             0
#> 196   57  Sepal.Width             2
#> 197   58 Petal.Length             1
#> 198   58  Petal.Width             1
#> 199   58 Sepal.Length             0
#> 200   58  Sepal.Width             2
#> 201   59 Petal.Length             0
#> 202   59  Petal.Width             1
#> 203   59 Sepal.Length             3
#> 204   59  Sepal.Width             4
#> 205   60 Petal.Length             2
#> 206   60  Petal.Width             0
#> 207   60 Sepal.Length             2
#> 208   60  Sepal.Width             3
#> 209   61 Petal.Length             0
#> 210   61  Petal.Width             3
#> 211   61 Sepal.Length             1
#> 212   62 Petal.Length             1
#> 213   62  Petal.Width             2
#> 214   62 Sepal.Length             0
#> 215   62  Sepal.Width             1
#> 216   63 Petal.Length             3
#> 217   63  Petal.Width             0
#> 218   63 Sepal.Length             3
#> 219   63  Sepal.Width             2
#> 220   64 Petal.Length             0
#> 221   64  Petal.Width             1
#> 222   64 Sepal.Length             2
#> 223   64  Sepal.Width             3
#> 224   65 Petal.Length             1
#> 225   65  Petal.Width             0
#> 226   65  Sepal.Width             4
#> 227   66 Petal.Length             1
#> 228   66  Petal.Width             0
#> 229   66 Sepal.Length             2
#> 230   67 Petal.Length             0
#> 231   67  Petal.Width             1
#> 232   68 Petal.Length             0
#> 233   68  Petal.Width             1
#> 234   68 Sepal.Length             2
#> 235   69 Petal.Length             1
#> 236   69  Petal.Width             2
#> 237   69 Sepal.Length             0
#> 238   69  Sepal.Width             1
#> 239   70 Petal.Length             2
#> 240   70  Petal.Width             0
#> 241   70 Sepal.Length             1
#> 242   71 Petal.Length             1
#> 243   71  Petal.Width             2
#> 244   71 Sepal.Length             0
#> 245   71  Sepal.Width             2
#> 246   72 Petal.Length             1
#> 247   72  Petal.Width             0
#> 248   72 Sepal.Length             3
#> 249   73 Petal.Length             2
#> 250   73  Petal.Width             0
#> 251   73 Sepal.Length             1
#> 252   73  Sepal.Width             3
#> 253   74 Petal.Length             1
#> 254   74  Petal.Width             0
#> 255   74 Sepal.Length             3
#> 256   75 Petal.Length             0
#> 257   75  Petal.Width             2
#> 258   75 Sepal.Length             3
#> 259   75  Sepal.Width             3
#> 260   76 Petal.Length             0
#> 261   76  Petal.Width             1
#> 262   76 Sepal.Length             3
#> 263   77 Petal.Length             0
#> 264   77  Petal.Width             2
#> 265   77  Sepal.Width             3
#> 266   78 Petal.Length             0
#> 267   78  Petal.Width             2
#> 268   78  Sepal.Width             3
#> 269   79 Petal.Length             1
#> 270   79  Petal.Width             2
#> 271   79 Sepal.Length             0
#> 272   79  Sepal.Width             1
#> 273   80 Petal.Length             0
#> 274   80  Petal.Width             3
#> 275   80 Sepal.Length             1
#> 276   80  Sepal.Width             5
#> 277   81 Petal.Length             1
#> 278   81  Petal.Width             2
#> 279   81 Sepal.Length             0
#> 280   81  Sepal.Width             2
#> 281   82 Petal.Length             3
#> 282   82  Petal.Width             1
#> 283   82 Sepal.Length             0
#> 284   82  Sepal.Width             1
#> 285   83 Petal.Length             1
#> 286   83  Petal.Width             0
#> 287   83 Sepal.Length             3
#> 288   83  Sepal.Width             4
#> 289   84 Petal.Length             0
#> 290   84  Petal.Width             2
#> 291   84 Sepal.Length             1
#> 292   84  Sepal.Width             5
#> 293   85 Petal.Length             1
#> 294   85  Petal.Width             1
#> 295   85 Sepal.Length             0
#> 296   86 Petal.Length             0
#> 297   86  Petal.Width             4
#> 298   86 Sepal.Length             2
#> 299   86  Sepal.Width             1
#> 300   87 Petal.Length             0
#> 301   87  Petal.Width             2
#> 302   87 Sepal.Length             4
#> 303   87  Sepal.Width             3
#> 304   88 Petal.Length             1
#> 305   88  Petal.Width             1
#> 306   88 Sepal.Length             0
#> 307   88  Sepal.Width             2
#> 308   89 Petal.Length             0
#> 309   89  Petal.Width             1
#> 310   89 Sepal.Length             4
#> 311   89  Sepal.Width             3
#> 312   90 Petal.Length             0
#> 313   90  Petal.Width             1
#> 314   90 Sepal.Length             2
#> 315   90  Sepal.Width             3
#> 316   91 Petal.Length             0
#> 317   91  Petal.Width             2
#> 318   91  Sepal.Width             3
#> 319   92 Petal.Length             0
#> 320   92  Petal.Width             5
#> 321   92 Sepal.Length             2
#> 322   92  Sepal.Width             3
#> 323   93 Petal.Length             2
#> 324   93  Petal.Width             0
#> 325   94 Petal.Length             0
#> 326   94  Petal.Width             2
#> 327   94 Sepal.Length             3
#> 328   94  Sepal.Width             4
#> 329   95 Petal.Length             0
#> 330   95  Petal.Width             1
#> 331   95 Sepal.Length             2
#> 332   96 Petal.Length             3
#> 333   96  Petal.Width             0
#>  [ reached 'max' / getOption("max.print") -- omitted 16 rows ]
min_depth_distribution(ranger::ranger(Species ~ ., data = iris, num.trees = 100))
#>     tree     variable minimal_depth
#> 1      1 Petal.Length             0
#> 2      1  Petal.Width             5
#> 3      1 Sepal.Length             3
#> 4      1  Sepal.Width             4
#> 5      2 Petal.Length             1
#> 6      2  Petal.Width             1
#> 7      2 Sepal.Length             0
#> 8      2  Sepal.Width             4
#> 9      3 Petal.Length             2
#> 10     3  Petal.Width             0
#> 11     3 Sepal.Length             7
#> 12     3  Sepal.Width             2
#> 13     4 Petal.Length             1
#> 14     4  Petal.Width             1
#> 15     4 Sepal.Length             0
#> 16     4  Sepal.Width             4
#> 17     5 Petal.Length             0
#> 18     5  Petal.Width             1
#> 19     5 Sepal.Length             2
#> 20     5  Sepal.Width             5
#> 21     6 Petal.Length             2
#> 22     6  Petal.Width             1
#> 23     6 Sepal.Length             0
#> 24     6  Sepal.Width             1
#> 25     7 Petal.Length             0
#> 26     7  Petal.Width             1
#> 27     7 Sepal.Length             4
#> 28     7  Sepal.Width             3
#> 29     8 Petal.Length             1
#> 30     8  Petal.Width             0
#> 31     8 Sepal.Length             2
#> 32     8  Sepal.Width             3
#> 33     9 Petal.Length             0
#> 34     9  Petal.Width             1
#> 35    10 Petal.Length             0
#> 36    10  Petal.Width             1
#> 37    10  Sepal.Width             3
#> 38    11 Petal.Length             1
#> 39    11  Petal.Width             0
#> 40    11  Sepal.Width             3
#> 41    12 Petal.Length             3
#> 42    12  Petal.Width             1
#> 43    12 Sepal.Length             0
#> 44    12  Sepal.Width             2
#> 45    13 Petal.Length             5
#> 46    13  Petal.Width             0
#> 47    13 Sepal.Length             1
#> 48    13  Sepal.Width             3
#> 49    14 Petal.Length             1
#> 50    14  Petal.Width             0
#> 51    14 Sepal.Length             2
#> 52    15 Petal.Length             0
#> 53    15  Petal.Width             1
#> 54    15 Sepal.Length             2
#> 55    16 Petal.Length             2
#> 56    16  Petal.Width             0
#> 57    16 Sepal.Length             2
#> 58    17 Petal.Length             1
#> 59    17  Petal.Width             3
#> 60    17 Sepal.Length             0
#> 61    17  Sepal.Width             2
#> 62    18 Petal.Length             0
#> 63    18  Petal.Width             2
#> 64    18 Sepal.Length             1
#> 65    18  Sepal.Width             4
#> 66    19 Petal.Length             0
#> 67    19  Petal.Width             1
#> 68    19  Sepal.Width             3
#> 69    20 Petal.Length             0
#> 70    20  Petal.Width             2
#> 71    20 Sepal.Length             3
#> 72    21 Petal.Length             0
#> 73    21  Petal.Width             1
#> 74    21 Sepal.Length             3
#> 75    21  Sepal.Width             3
#> 76    22 Petal.Length             0
#> 77    22  Petal.Width             2
#> 78    22 Sepal.Length             3
#> 79    22  Sepal.Width             2
#> 80    23 Petal.Length             1
#> 81    23  Petal.Width             0
#> 82    23 Sepal.Length             3
#> 83    23  Sepal.Width             3
#> 84    24 Petal.Length             1
#> 85    24  Petal.Width             2
#> 86    24 Sepal.Length             0
#> 87    24  Sepal.Width             1
#> 88    25 Petal.Length             1
#> 89    25  Petal.Width             1
#> 90    25 Sepal.Length             0
#> 91    26 Petal.Length             0
#> 92    26  Petal.Width             1
#> 93    26 Sepal.Length             4
#> 94    27 Petal.Length             0
#> 95    27  Petal.Width             2
#> 96    27 Sepal.Length             3
#> 97    27  Sepal.Width             3
#> 98    28 Petal.Length             0
#> 99    28  Petal.Width             1
#> 100   28 Sepal.Length             2
#> 101   28  Sepal.Width             2
#> 102   29 Petal.Length             1
#> 103   29  Petal.Width             0
#> 104   30 Petal.Length             1
#> 105   30  Petal.Width             0
#> 106   30 Sepal.Length             3
#> 107   31 Petal.Length             0
#> 108   31  Petal.Width             1
#> 109   31 Sepal.Length             2
#> 110   32 Petal.Length             1
#> 111   32  Petal.Width             0
#> 112   32 Sepal.Length             3
#> 113   33 Petal.Length             0
#> 114   33  Petal.Width             1
#> 115   33 Sepal.Length             3
#> 116   33  Sepal.Width             3
#> 117   34 Petal.Length             0
#> 118   34  Petal.Width             1
#> 119   34 Sepal.Length             2
#> 120   34  Sepal.Width             3
#> 121   35 Petal.Length             3
#> 122   35  Petal.Width             0
#> 123   35 Sepal.Length             2
#> 124   36 Petal.Length             2
#> 125   36  Petal.Width             0
#> 126   36 Sepal.Length             2
#> 127   36  Sepal.Width             3
#> 128   37 Petal.Length             1
#> 129   37  Petal.Width             3
#> 130   37 Sepal.Length             0
#> 131   37  Sepal.Width             1
#> 132   38 Petal.Length             1
#> 133   38  Petal.Width             0
#> 134   38 Sepal.Length             3
#> 135   38  Sepal.Width             4
#> 136   39 Petal.Length             1
#> 137   39  Petal.Width             0
#> 138   39 Sepal.Length             4
#> 139   39  Sepal.Width             3
#> 140   40 Petal.Length             0
#> 141   40  Petal.Width             1
#> 142   40 Sepal.Length             3
#> 143   40  Sepal.Width             3
#> 144   41 Petal.Length             1
#> 145   41  Petal.Width             0
#> 146   41 Sepal.Length             3
#> 147   41  Sepal.Width             3
#> 148   42 Petal.Length             0
#> 149   42  Petal.Width             2
#> 150   42 Sepal.Length             1
#> 151   42  Sepal.Width             3
#> 152   43 Petal.Length             0
#> 153   43  Petal.Width             2
#> 154   43 Sepal.Length             2
#> 155   43  Sepal.Width             4
#> 156   44 Petal.Length             1
#> 157   44  Petal.Width             0
#> 158   44 Sepal.Length             2
#> 159   44  Sepal.Width             3
#> 160   45 Petal.Length             0
#> 161   45  Petal.Width             2
#> 162   45 Sepal.Length             3
#> 163   45  Sepal.Width             4
#> 164   46 Petal.Length             2
#> 165   46  Petal.Width             0
#> 166   46 Sepal.Length             1
#> 167   46  Sepal.Width             3
#> 168   47 Petal.Length             2
#> 169   47  Petal.Width             0
#> 170   47 Sepal.Length             2
#> 171   47  Sepal.Width             4
#> 172   48 Petal.Length             0
#> 173   48  Petal.Width             2
#> 174   49 Petal.Length             1
#> 175   49  Petal.Width             0
#> 176   49 Sepal.Length             3
#> 177   49  Sepal.Width             4
#> 178   50 Petal.Length             4
#> 179   50  Petal.Width             0
#> 180   50 Sepal.Length             2
#> 181   50  Sepal.Width             2
#> 182   51 Petal.Length             1
#> 183   51  Petal.Width             2
#> 184   51 Sepal.Length             0
#> 185   52 Petal.Length             1
#> 186   52  Petal.Width             0
#> 187   52 Sepal.Length             3
#> 188   52  Sepal.Width             2
#> 189   53 Petal.Length             0
#> 190   53  Petal.Width             1
#> 191   53 Sepal.Length             2
#> 192   53  Sepal.Width             3
#> 193   54 Petal.Length             0
#> 194   54  Petal.Width             3
#> 195   54 Sepal.Length             3
#> 196   54  Sepal.Width             2
#> 197   55 Petal.Length             0
#> 198   55  Petal.Width             1
#> 199   55 Sepal.Length             2
#> 200   56 Petal.Length             1
#> 201   56  Petal.Width             1
#> 202   56 Sepal.Length             0
#> 203   57 Petal.Length             2
#> 204   57  Petal.Width             0
#> 205   57 Sepal.Length             2
#> 206   57  Sepal.Width             3
#> 207   58 Petal.Length             2
#> 208   58  Petal.Width             1
#> 209   58 Sepal.Length             0
#> 210   59 Petal.Length             0
#> 211   59  Petal.Width             1
#> 212   59 Sepal.Length             2
#> 213   59  Sepal.Width             4
#> 214   60 Petal.Length             3
#> 215   60  Petal.Width             0
#> 216   60 Sepal.Length             1
#> 217   60  Sepal.Width             4
#> 218   61 Petal.Length             0
#> 219   61  Petal.Width             2
#> 220   61 Sepal.Length             1
#> 221   61  Sepal.Width             3
#> 222   62 Petal.Length             2
#> 223   62  Petal.Width             0
#> 224   62 Sepal.Length             1
#> 225   62  Sepal.Width             3
#> 226   63 Petal.Length             2
#> 227   63  Petal.Width             0
#> 228   63 Sepal.Length             3
#> 229   63  Sepal.Width             3
#> 230   64 Petal.Length             1
#> 231   64  Petal.Width             0
#> 232   65 Petal.Length             0
#> 233   65  Petal.Width             2
#> 234   65 Sepal.Length             2
#> 235   66 Petal.Length             2
#> 236   66  Petal.Width             1
#> 237   66 Sepal.Length             0
#> 238   66  Sepal.Width             1
#> 239   67  Petal.Width             0
#> 240   67 Sepal.Length             1
#> 241   67  Sepal.Width             4
#> 242   68 Petal.Length             0
#> 243   68  Petal.Width             2
#> 244   68 Sepal.Length             2
#> 245   68  Sepal.Width             3
#> 246   69 Petal.Length             3
#> 247   69  Petal.Width             0
#> 248   69 Sepal.Length             1
#> 249   70 Petal.Length             2
#> 250   70  Petal.Width             0
#> 251   70  Sepal.Width             4
#> 252   71 Petal.Length             0
#> 253   71  Petal.Width             1
#> 254   71 Sepal.Length             2
#> 255   71  Sepal.Width             3
#> 256   72 Petal.Length             2
#> 257   72  Petal.Width             0
#> 258   72 Sepal.Length             3
#> 259   73 Petal.Length             0
#> 260   73 Sepal.Length             2
#> 261   73  Sepal.Width             5
#> 262   74 Petal.Length             2
#> 263   74  Petal.Width             0
#> 264   74 Sepal.Length             4
#> 265   75 Petal.Length             0
#> 266   75  Petal.Width             1
#> 267   75 Sepal.Length             2
#> 268   75  Sepal.Width             2
#> 269   76 Petal.Length             1
#> 270   76  Petal.Width             0
#> 271   76 Sepal.Length             3
#> 272   76  Sepal.Width             4
#> 273   77 Petal.Length             2
#> 274   77  Petal.Width             0
#> 275   77 Sepal.Length             1
#> 276   77  Sepal.Width             3
#> 277   78 Petal.Length             0
#> 278   78  Petal.Width             1
#> 279   78 Sepal.Length             2
#> 280   79 Petal.Length             1
#> 281   79  Petal.Width             0
#> 282   79 Sepal.Length             4
#> 283   79  Sepal.Width             4
#> 284   80 Petal.Length             0
#> 285   80  Petal.Width             2
#> 286   80 Sepal.Length             4
#> 287   80  Sepal.Width             2
#> 288   81 Petal.Length             1
#> 289   81  Petal.Width             0
#> 290   81 Sepal.Length             2
#> 291   81  Sepal.Width             4
#> 292   82 Petal.Length             2
#> 293   82  Petal.Width             0
#> 294   82 Sepal.Length             2
#> 295   82  Sepal.Width             3
#> 296   83 Petal.Length             1
#> 297   83  Petal.Width             1
#> 298   83 Sepal.Length             0
#> 299   84 Petal.Length             0
#> 300   84  Petal.Width             4
#> 301   84 Sepal.Length             2
#> 302   84  Sepal.Width             5
#> 303   85 Petal.Length             3
#> 304   85  Petal.Width             0
#> 305   85 Sepal.Length             1
#> 306   85  Sepal.Width             4
#> 307   86 Petal.Length             2
#> 308   86  Petal.Width             0
#> 309   86 Sepal.Length             3
#> 310   87 Petal.Length             0
#> 311   87  Petal.Width             2
#> 312   87 Sepal.Length             4
#> 313   87  Sepal.Width             3
#> 314   88 Petal.Length             0
#> 315   88  Petal.Width             1
#> 316   88 Sepal.Length             2
#> 317   88  Sepal.Width             3
#> 318   89 Petal.Length             0
#> 319   89  Petal.Width             1
#> 320   89 Sepal.Length             4
#> 321   89  Sepal.Width             3
#> 322   90 Petal.Length             1
#> 323   90  Petal.Width             1
#> 324   90 Sepal.Length             2
#> 325   90  Sepal.Width             0
#> 326   91 Petal.Length             1
#> 327   91  Petal.Width             0
#> 328   92 Petal.Length             0
#> 329   92  Petal.Width             2
#> 330   92  Sepal.Width             6
#> 331   93 Petal.Length             0
#> 332   93  Petal.Width             1
#> 333   93 Sepal.Length             2
#>  [ reached 'max' / getOption("max.print") -- omitted 25 rows ]