Calculate minimal depth distribution of a random forest

Get minimal depth values for all trees in a random forest

min_depth_distribution(forest)

Arguments

forest

A randomForest or ranger object

Value

A data frame with the value of minimal depth for every variable in every tree

Examples

min_depth_distribution(randomForest::randomForest(Species ~ ., data = iris, ntree = 100))
#>     tree     variable minimal_depth
#> 1      1 Petal.Length             1
#> 2      1  Petal.Width             2
#> 3      1 Sepal.Length             0
#> 4      2 Petal.Length             0
#> 5      2  Petal.Width             1
#> 6      2 Sepal.Length             3
#> 7      2  Sepal.Width             2
#> 8      3 Petal.Length             0
#> 9      3  Petal.Width             3
#> 10     3 Sepal.Length             1
#> 11     3  Sepal.Width             3
#> 12     4 Petal.Length             0
#> 13     4  Petal.Width             4
#> 14     4 Sepal.Length             3
#> 15     4  Sepal.Width             3
#> 16     5 Petal.Length             0
#> 17     5  Petal.Width             2
#> 18     5 Sepal.Length             1
#> 19     6 Petal.Length             3
#> 20     6  Petal.Width             0
#> 21     6 Sepal.Length             1
#> 22     7 Petal.Length             2
#> 23     7  Petal.Width             0
#> 24     8 Petal.Length             0
#> 25     8  Petal.Width             1
#> 26     8 Sepal.Length             2
#> 27     9 Petal.Length             3
#> 28     9  Petal.Width             0
#> 29     9 Sepal.Length             4
#> 30     9  Sepal.Width             2
#> 31    10 Petal.Length             3
#> 32    10  Petal.Width             0
#> 33    10 Sepal.Length             2
#> 34    10  Sepal.Width             3
#> 35    11 Petal.Length             0
#> 36    11  Petal.Width             1
#> 37    11 Sepal.Length             5
#> 38    12 Petal.Length             0
#> 39    12  Petal.Width             1
#> 40    12 Sepal.Length             2
#> 41    12  Sepal.Width             3
#> 42    13 Petal.Length             4
#> 43    13  Petal.Width             0
#> 44    13 Sepal.Length             1
#> 45    13  Sepal.Width             4
#> 46    14 Petal.Length             0
#> 47    14  Petal.Width             2
#> 48    14 Sepal.Length             3
#> 49    14  Sepal.Width             2
#> 50    15 Petal.Length             1
#> 51    15  Petal.Width             0
#> 52    15 Sepal.Length             2
#> 53    16 Petal.Length             0
#> 54    16  Petal.Width             1
#> 55    17 Petal.Length             0
#> 56    17  Petal.Width             2
#> 57    17 Sepal.Length             2
#> 58    18 Petal.Length             1
#> 59    18  Petal.Width             1
#> 60    18 Sepal.Length             0
#> 61    18  Sepal.Width             2
#> 62    19 Petal.Length             0
#> 63    19  Petal.Width             3
#> 64    19 Sepal.Length             2
#> 65    19  Sepal.Width             2
#> 66    20 Petal.Length             4
#> 67    20  Petal.Width             0
#> 68    20  Sepal.Width             3
#> 69    21  Petal.Width             0
#> 70    21 Sepal.Length             2
#> 71    21  Sepal.Width             3
#> 72    22 Petal.Length             1
#> 73    22  Petal.Width             2
#> 74    22 Sepal.Length             0
#> 75    22  Sepal.Width             1
#> 76    23 Petal.Length             0
#> 77    23  Petal.Width             2
#> 78    23 Sepal.Length             1
#> 79    23  Sepal.Width             2
#> 80    24 Petal.Length             1
#> 81    24  Petal.Width             1
#> 82    24 Sepal.Length             0
#> 83    24  Sepal.Width             3
#> 84    25 Petal.Length             0
#> 85    25  Petal.Width             2
#> 86    25 Sepal.Length             2
#> 87    25  Sepal.Width             4
#> 88    26 Petal.Length             0
#> 89    26  Petal.Width             2
#> 90    26 Sepal.Length             5
#> 91    26  Sepal.Width             2
#> 92    27 Petal.Length             0
#> 93    27  Petal.Width             3
#> 94    27 Sepal.Length             1
#> 95    27  Sepal.Width             3
#> 96    28 Petal.Length             2
#> 97    28  Petal.Width             0
#> 98    28 Sepal.Length             1
#> 99    29 Petal.Length             1
#> 100   29  Petal.Width             4
#> 101   29 Sepal.Length             0
#> 102   29  Sepal.Width             3
#> 103   30 Petal.Length             0
#> 104   30  Petal.Width             2
#> 105   30 Sepal.Length             1
#> 106   30  Sepal.Width             5
#> 107   31 Petal.Length             0
#> 108   31  Petal.Width             1
#> 109   31  Sepal.Width             2
#> 110   32 Petal.Length             0
#> 111   32  Petal.Width             2
#> 112   32 Sepal.Length             1
#> 113   32  Sepal.Width             1
#> 114   33 Petal.Length             1
#> 115   33  Petal.Width             1
#> 116   33 Sepal.Length             0
#> 117   34 Petal.Length             1
#> 118   34  Petal.Width             2
#> 119   34 Sepal.Length             0
#> 120   34  Sepal.Width             4
#> 121   35 Petal.Length             0
#> 122   35  Petal.Width             1
#> 123   35 Sepal.Length             2
#> 124   36 Petal.Length             1
#> 125   36  Petal.Width             1
#> 126   36 Sepal.Length             0
#> 127   37 Petal.Length             2
#> 128   37  Petal.Width             0
#> 129   37 Sepal.Length             2
#> 130   37  Sepal.Width             3
#> 131   38 Petal.Length             0
#> 132   38  Petal.Width             1
#> 133   38 Sepal.Length             4
#> 134   39  Petal.Width             0
#> 135   39 Sepal.Length             4
#> 136   39  Sepal.Width             2
#> 137   40 Petal.Length             1
#> 138   40  Petal.Width             0
#> 139   40 Sepal.Length             2
#> 140   40  Sepal.Width             3
#> 141   41 Petal.Length             0
#> 142   41  Petal.Width             2
#> 143   41 Sepal.Length             1
#> 144   42 Petal.Length             1
#> 145   42  Petal.Width             1
#> 146   42 Sepal.Length             0
#> 147   42  Sepal.Width             4
#> 148   43 Petal.Length             3
#> 149   43  Petal.Width             0
#> 150   43 Sepal.Length             1
#> 151   43  Sepal.Width             4
#> 152   44 Petal.Length             0
#> 153   44  Petal.Width             2
#> 154   44 Sepal.Length             5
#> 155   45 Petal.Length             0
#> 156   45  Petal.Width             2
#> 157   45 Sepal.Length             2
#> 158   45  Sepal.Width             3
#> 159   46 Petal.Length             2
#> 160   46  Petal.Width             1
#> 161   46 Sepal.Length             0
#> 162   46  Sepal.Width             3
#> 163   47 Petal.Length             2
#> 164   47  Petal.Width             0
#> 165   47 Sepal.Length             3
#> 166   48 Petal.Length             2
#> 167   48  Petal.Width             0
#> 168   48 Sepal.Length             2
#> 169   48  Sepal.Width             5
#> 170   49 Petal.Length             3
#> 171   49  Petal.Width             1
#> 172   49 Sepal.Length             0
#> 173   49  Sepal.Width             2
#> 174   50 Petal.Length             0
#> 175   50  Petal.Width             1
#> 176   51 Petal.Length             1
#> 177   51  Petal.Width             1
#> 178   51 Sepal.Length             0
#> 179   51  Sepal.Width             3
#> 180   52 Petal.Length             2
#> 181   52  Petal.Width             0
#> 182   52 Sepal.Length             2
#> 183   53 Petal.Length             0
#> 184   53  Petal.Width             1
#> 185   53 Sepal.Length             2
#> 186   53  Sepal.Width             3
#> 187   54 Petal.Length             0
#> 188   54  Petal.Width             1
#> 189   54 Sepal.Length             3
#> 190   54  Sepal.Width             3
#> 191   55 Petal.Length             2
#> 192   55  Petal.Width             0
#> 193   55 Sepal.Length             1
#> 194   55  Sepal.Width             4
#> 195   56 Petal.Length             0
#> 196   56  Petal.Width             2
#> 197   56 Sepal.Length             4
#> 198   57 Petal.Length             1
#> 199   57  Petal.Width             1
#> 200   57 Sepal.Length             0
#> 201   57  Sepal.Width             3
#> 202   58 Petal.Length             1
#> 203   58  Petal.Width             1
#> 204   58 Sepal.Length             0
#> 205   59 Petal.Length             1
#> 206   59  Petal.Width             0
#> 207   59  Sepal.Width             3
#> 208   60 Petal.Length             0
#> 209   60  Petal.Width             1
#> 210   60 Sepal.Length             4
#> 211   60  Sepal.Width             3
#> 212   61 Petal.Length             0
#> 213   61  Petal.Width             1
#> 214   61 Sepal.Length             2
#> 215   62 Petal.Length             0
#> 216   62  Petal.Width             2
#> 217   62 Sepal.Length             1
#> 218   62  Sepal.Width             4
#> 219   63 Petal.Length             2
#> 220   63  Petal.Width             0
#> 221   63 Sepal.Length             2
#> 222   63  Sepal.Width             3
#> 223   64 Petal.Length             0
#> 224   64  Petal.Width             2
#> 225   64 Sepal.Length             3
#> 226   64  Sepal.Width             3
#> 227   65 Petal.Length             1
#> 228   65  Petal.Width             2
#> 229   65 Sepal.Length             0
#> 230   66 Petal.Length             0
#> 231   66  Petal.Width             2
#> 232   66 Sepal.Length             3
#> 233   67 Petal.Length             2
#> 234   67  Petal.Width             3
#> 235   67 Sepal.Length             0
#> 236   67  Sepal.Width             1
#> 237   68 Petal.Length             5
#> 238   68  Petal.Width             1
#> 239   68 Sepal.Length             0
#> 240   68  Sepal.Width             2
#> 241   69 Petal.Length             0
#> 242   69  Petal.Width             3
#> 243   69 Sepal.Length             1
#> 244   69  Sepal.Width             6
#> 245   70 Petal.Length             1
#> 246   70  Petal.Width             0
#> 247   70 Sepal.Length             3
#> 248   70  Sepal.Width             4
#> 249   71 Petal.Length             3
#> 250   71  Petal.Width             0
#> 251   71 Sepal.Length             3
#> 252   72 Petal.Length             0
#> 253   72  Petal.Width             3
#> 254   72 Sepal.Length             3
#> 255   72  Sepal.Width             4
#> 256   73 Petal.Length             0
#> 257   73  Petal.Width             1
#> 258   73 Sepal.Length             4
#> 259   74 Petal.Length             0
#> 260   74  Petal.Width             2
#> 261   74 Sepal.Length             3
#> 262   74  Sepal.Width             4
#> 263   75 Petal.Length             2
#> 264   75  Petal.Width             0
#> 265   75 Sepal.Length             1
#> 266   76 Petal.Length             0
#> 267   76  Petal.Width             1
#> 268   76 Sepal.Length             2
#> 269   77 Petal.Length             2
#> 270   77  Petal.Width             0
#> 271   77 Sepal.Length             4
#> 272   77  Sepal.Width             3
#> 273   78 Petal.Length             3
#> 274   78  Petal.Width             0
#> 275   78 Sepal.Length             2
#> 276   78  Sepal.Width             2
#> 277   79 Petal.Length             1
#> 278   79  Petal.Width             2
#> 279   79 Sepal.Length             0
#> 280   80 Petal.Length             1
#> 281   80  Petal.Width             1
#> 282   80 Sepal.Length             0
#> 283   80  Sepal.Width             3
#> 284   81 Petal.Length             3
#> 285   81  Petal.Width             0
#> 286   81 Sepal.Length             2
#> 287   81  Sepal.Width             4
#> 288   82 Petal.Length             0
#> 289   82  Petal.Width             3
#> 290   82 Sepal.Length             3
#> 291   82  Sepal.Width             2
#> 292   83 Petal.Length             0
#> 293   83  Petal.Width             3
#> 294   83 Sepal.Length             1
#> 295   84 Petal.Length             3
#> 296   84  Petal.Width             0
#> 297   84  Sepal.Width             2
#> 298   85 Petal.Length             0
#> 299   85  Petal.Width             1
#> 300   85 Sepal.Length             2
#> 301   86 Petal.Length             0
#> 302   86  Petal.Width             1
#> 303   87 Petal.Length             0
#> 304   87  Petal.Width             1
#> 305   87 Sepal.Length             2
#> 306   88 Petal.Length             0
#> 307   88  Petal.Width             1
#> 308   88 Sepal.Length             4
#> 309   89 Petal.Length             0
#> 310   89  Petal.Width             1
#> 311   89 Sepal.Length             3
#> 312   90 Petal.Length             1
#> 313   90  Petal.Width             1
#> 314   90 Sepal.Length             0
#> 315   90  Sepal.Width             2
#> 316   91 Petal.Length             1
#> 317   91  Petal.Width             0
#> 318   91 Sepal.Length             4
#> 319   92 Petal.Length             4
#> 320   92  Petal.Width             0
#> 321   92 Sepal.Length             1
#> 322   93 Petal.Length             1
#> 323   93  Petal.Width             0
#> 324   94 Petal.Length             2
#> 325   94  Petal.Width             0
#> 326   94 Sepal.Length             1
#> 327   94  Sepal.Width             2
#> 328   95 Petal.Length             1
#> 329   95  Petal.Width             0
#> 330   95  Sepal.Width             2
#> 331   96 Petal.Length             0
#> 332   96  Petal.Width             3
#> 333   96 Sepal.Length             2
#> 334   97 Petal.Length             1
#> 335   97  Petal.Width             0
#> 336   97 Sepal.Length             2
#> 337   97  Sepal.Width             3
#> 338   98 Petal.Length             2
#> 339   98  Petal.Width             0
#> 340   98 Sepal.Length             3
#> 341   99 Petal.Length             3
#> 342   99  Petal.Width             0
#> 343   99 Sepal.Length             1
#> 344   99  Sepal.Width             4
#> 345  100 Petal.Length             1
#> 346  100  Petal.Width             0
#> 347  100 Sepal.Length             3
min_depth_distribution(ranger::ranger(Species ~ ., data = iris, num.trees = 100))
#>     tree     variable minimal_depth
#> 1      1 Petal.Length             0
#> 2      1  Petal.Width             1
#> 3      1 Sepal.Length             2
#> 4      1  Sepal.Width             4
#> 5      2 Petal.Length             0
#> 6      2  Petal.Width             1
#> 7      2 Sepal.Length             3
#> 8      2  Sepal.Width             2
#> 9      3 Petal.Length             0
#> 10     3  Petal.Width             3
#> 11     3 Sepal.Length             1
#> 12     3  Sepal.Width             3
#> 13     4 Petal.Length             0
#> 14     4  Petal.Width             3
#> 15     4 Sepal.Length             4
#> 16     4  Sepal.Width             2
#> 17     5 Petal.Length             0
#> 18     5  Petal.Width             3
#> 19     5 Sepal.Length             2
#> 20     5  Sepal.Width             1
#> 21     6 Petal.Length             1
#> 22     6  Petal.Width             2
#> 23     6 Sepal.Length             0
#> 24     6  Sepal.Width             1
#> 25     7 Petal.Length             1
#> 26     7  Petal.Width             0
#> 27     7 Sepal.Length             2
#> 28     7  Sepal.Width             5
#> 29     8 Petal.Length             2
#> 30     8  Petal.Width             0
#> 31     8 Sepal.Length             2
#> 32     8  Sepal.Width             4
#> 33     9 Petal.Length             1
#> 34     9  Petal.Width             0
#> 35    10 Petal.Length             1
#> 36    10  Petal.Width             0
#> 37    10  Sepal.Width             3
#> 38    11 Petal.Length             0
#> 39    11  Petal.Width             5
#> 40    11 Sepal.Length             2
#> 41    11  Sepal.Width             3
#> 42    12 Petal.Length             0
#> 43    12  Petal.Width             2
#> 44    12  Sepal.Width             3
#> 45    13 Petal.Length             1
#> 46    13  Petal.Width             0
#> 47    13 Sepal.Length             3
#> 48    14 Petal.Length             1
#> 49    14  Petal.Width             0
#> 50    14 Sepal.Length             2
#> 51    14  Sepal.Width             3
#> 52    15 Petal.Length             2
#> 53    15  Petal.Width             0
#> 54    15 Sepal.Length             3
#> 55    16 Petal.Length             0
#> 56    16  Petal.Width             2
#> 57    16 Sepal.Length             2
#> 58    17 Petal.Length             0
#> 59    17  Petal.Width             1
#> 60    17 Sepal.Length             2
#> 61    17  Sepal.Width             3
#> 62    18  Petal.Width             1
#> 63    18 Sepal.Length             0
#> 64    18  Sepal.Width             2
#> 65    19 Petal.Length             0
#> 66    19  Petal.Width             2
#> 67    19 Sepal.Length             4
#> 68    19  Sepal.Width             3
#> 69    20 Petal.Length             0
#> 70    20  Petal.Width             1
#> 71    21 Petal.Length             1
#> 72    21  Petal.Width             1
#> 73    21 Sepal.Length             0
#> 74    21  Sepal.Width             5
#> 75    22 Petal.Length             0
#> 76    22  Petal.Width             1
#> 77    22 Sepal.Length             3
#> 78    23 Petal.Length             1
#> 79    23  Petal.Width             0
#> 80    23  Sepal.Width             3
#> 81    24 Petal.Length             0
#> 82    24  Petal.Width             4
#> 83    24 Sepal.Length             3
#> 84    24  Sepal.Width             3
#> 85    25 Petal.Length             1
#> 86    25  Petal.Width             1
#> 87    25 Sepal.Length             0
#> 88    25  Sepal.Width             3
#> 89    26 Petal.Length             1
#> 90    26  Petal.Width             0
#> 91    26 Sepal.Length             2
#> 92    27 Petal.Length             2
#> 93    27  Petal.Width             0
#> 94    27 Sepal.Length             1
#> 95    27  Sepal.Width             5
#> 96    28 Petal.Length             0
#> 97    28  Petal.Width             2
#> 98    29 Petal.Length             2
#> 99    29  Petal.Width             0
#> 100   29 Sepal.Length             4
#> 101   30 Petal.Length             2
#> 102   30  Petal.Width             0
#> 103   30 Sepal.Length             1
#> 104   30  Sepal.Width             3
#> 105   31 Petal.Length             0
#> 106   31  Petal.Width             3
#> 107   31 Sepal.Length             2
#> 108   31  Sepal.Width             5
#> 109   32 Petal.Length             1
#> 110   32  Petal.Width             0
#> 111   32 Sepal.Length             2
#> 112   32  Sepal.Width             3
#> 113   33 Petal.Length             2
#> 114   33  Petal.Width             0
#> 115   33 Sepal.Length             1
#> 116   33  Sepal.Width             1
#> 117   34 Petal.Length             3
#> 118   34  Petal.Width             0
#> 119   34 Sepal.Length             3
#> 120   34  Sepal.Width             2
#> 121   35 Petal.Length             2
#> 122   35  Petal.Width             0
#> 123   35  Sepal.Width             3
#> 124   36 Petal.Length             1
#> 125   36  Petal.Width             2
#> 126   36 Sepal.Length             0
#> 127   36  Sepal.Width             3
#> 128   37 Petal.Length             0
#> 129   37  Petal.Width             1
#> 130   37  Sepal.Width             4
#> 131   38 Petal.Length             0
#> 132   38  Petal.Width             2
#> 133   38 Sepal.Length             1
#> 134   38  Sepal.Width             3
#> 135   39 Petal.Length             2
#> 136   39  Petal.Width             0
#> 137   39 Sepal.Length             1
#> 138   40 Petal.Length             0
#> 139   40  Petal.Width             3
#> 140   40 Sepal.Length             1
#> 141   40  Sepal.Width             3
#> 142   41 Petal.Length             3
#> 143   41  Petal.Width             0
#> 144   41 Sepal.Length             2
#> 145   42 Petal.Length             2
#> 146   42  Petal.Width             1
#> 147   42 Sepal.Length             0
#> 148   42  Sepal.Width             1
#> 149   43 Petal.Length             1
#> 150   43  Petal.Width             2
#> 151   43 Sepal.Length             0
#> 152   44 Petal.Length             3
#> 153   44  Petal.Width             1
#> 154   44 Sepal.Length             0
#> 155   44  Sepal.Width             2
#> 156   45 Petal.Length             1
#> 157   45  Petal.Width             1
#> 158   45 Sepal.Length             0
#> 159   45  Sepal.Width             3
#> 160   46 Petal.Length             2
#> 161   46  Petal.Width             1
#> 162   46 Sepal.Length             0
#> 163   46  Sepal.Width             3
#> 164   47 Petal.Length             0
#> 165   47  Petal.Width             1
#> 166   47 Sepal.Length             4
#> 167   47  Sepal.Width             3
#> 168   48 Petal.Length             0
#> 169   48  Petal.Width             2
#> 170   48 Sepal.Length             1
#> 171   49 Petal.Length             1
#> 172   49  Petal.Width             1
#> 173   49 Sepal.Length             0
#> 174   50 Petal.Length             1
#> 175   50  Petal.Width             0
#> 176   50  Sepal.Width             2
#> 177   51 Petal.Length             1
#> 178   51  Petal.Width             1
#> 179   51 Sepal.Length             0
#> 180   52 Petal.Length             0
#> 181   52  Petal.Width             1
#> 182   52 Sepal.Length             6
#> 183   52  Sepal.Width             2
#> 184   53 Petal.Length             1
#> 185   53  Petal.Width             0
#> 186   53 Sepal.Length             2
#> 187   53  Sepal.Width             3
#> 188   54 Petal.Length             0
#> 189   54  Petal.Width             2
#> 190   54  Sepal.Width             2
#> 191   55 Petal.Length             1
#> 192   55  Petal.Width             2
#> 193   55 Sepal.Length             0
#> 194   56 Petal.Length             2
#> 195   56  Petal.Width             0
#> 196   56 Sepal.Length             1
#> 197   56  Sepal.Width             3
#> 198   57 Petal.Length             0
#> 199   57  Petal.Width             2
#> 200   57 Sepal.Length             1
#> 201   58 Petal.Length             0
#> 202   58  Petal.Width             2
#> 203   58 Sepal.Length             3
#> 204   58  Sepal.Width             5
#> 205   59 Petal.Length             2
#> 206   59  Petal.Width             0
#> 207   59 Sepal.Length             2
#> 208   59  Sepal.Width             4
#> 209   60 Petal.Length             0
#> 210   60  Petal.Width             2
#> 211   60 Sepal.Length             1
#> 212   61 Petal.Length             2
#> 213   61  Petal.Width             2
#> 214   61 Sepal.Length             0
#> 215   61  Sepal.Width             1
#> 216   62 Petal.Length             1
#> 217   62  Petal.Width             4
#> 218   62 Sepal.Length             0
#> 219   62  Sepal.Width             3
#> 220   63 Petal.Length             2
#> 221   63  Petal.Width             0
#> 222   63  Sepal.Width             3
#> 223   64 Petal.Length             0
#> 224   64  Petal.Width             1
#> 225   64 Sepal.Length             2
#> 226   64  Sepal.Width             3
#> 227   65 Petal.Length             0
#> 228   65  Petal.Width             1
#> 229   65  Sepal.Width             2
#> 230   66 Petal.Length             1
#> 231   66  Petal.Width             0
#> 232   66  Sepal.Width             3
#> 233   67 Petal.Length             2
#> 234   67  Petal.Width             1
#> 235   67 Sepal.Length             0
#> 236   67  Sepal.Width             1
#> 237   68 Petal.Length             0
#> 238   68 Sepal.Length             3
#> 239   69 Petal.Length             2
#> 240   69  Petal.Width             0
#> 241   69 Sepal.Length             5
#> 242   69  Sepal.Width             2
#> 243   70 Petal.Length             1
#> 244   70  Petal.Width             2
#> 245   70 Sepal.Length             0
#> 246   70  Sepal.Width             1
#> 247   71 Petal.Length             3
#> 248   71  Petal.Width             0
#> 249   71 Sepal.Length             1
#> 250   71  Sepal.Width             4
#> 251   72 Petal.Length             1
#> 252   72  Petal.Width             1
#> 253   72 Sepal.Length             0
#> 254   72  Sepal.Width             2
#> 255   73 Petal.Length             1
#> 256   73  Petal.Width             0
#> 257   73 Sepal.Length             3
#> 258   73  Sepal.Width             4
#> 259   74 Petal.Length             2
#> 260   74  Petal.Width             0
#> 261   74 Sepal.Length             2
#> 262   74  Sepal.Width             5
#> 263   75 Petal.Length             3
#> 264   75  Petal.Width             0
#> 265   75 Sepal.Length             1
#> 266   76 Petal.Length             2
#> 267   76  Petal.Width             0
#> 268   76 Sepal.Length             3
#> 269   77 Petal.Length             0
#> 270   77  Petal.Width             2
#> 271   77 Sepal.Length             1
#> 272   77  Sepal.Width             5
#> 273   78 Petal.Length             2
#> 274   78  Petal.Width             0
#> 275   78  Sepal.Width             3
#> 276   79 Petal.Length             1
#> 277   79  Petal.Width             0
#> 278   79 Sepal.Length             3
#> 279   79  Sepal.Width             4
#> 280   80 Petal.Length             3
#> 281   80  Petal.Width             0
#> 282   80 Sepal.Length             3
#> 283   80  Sepal.Width             2
#> 284   81 Petal.Length             1
#> 285   81  Petal.Width             0
#> 286   81 Sepal.Length             4
#> 287   81  Sepal.Width             2
#> 288   82 Petal.Length             0
#> 289   82  Petal.Width             1
#> 290   82 Sepal.Length             2
#> 291   83 Petal.Length             1
#> 292   83  Petal.Width             0
#> 293   83 Sepal.Length             2
#> 294   83  Sepal.Width             4
#> 295   84 Petal.Length             0
#> 296   84  Petal.Width             1
#> 297   84 Sepal.Length             2
#> 298   84  Sepal.Width             3
#> 299   85 Petal.Length             1
#> 300   85  Petal.Width             0
#> 301   85 Sepal.Length             3
#> 302   86 Petal.Length             1
#> 303   86  Petal.Width             0
#> 304   86 Sepal.Length             2
#> 305   87 Petal.Length             0
#> 306   87  Petal.Width             1
#> 307   87  Sepal.Width             3
#> 308   88 Petal.Length             0
#> 309   88  Petal.Width             3
#> 310   88 Sepal.Length             3
#> 311   88  Sepal.Width             2
#> 312   89 Petal.Length             3
#> 313   89  Petal.Width             0
#> 314   89 Sepal.Length             1
#> 315   89  Sepal.Width             3
#> 316   90 Petal.Length             2
#> 317   90  Petal.Width             0
#> 318   90 Sepal.Length             1
#> 319   91 Petal.Length             1
#> 320   91  Petal.Width             0
#> 321   91 Sepal.Length             3
#> 322   91  Sepal.Width             4
#> 323   92 Petal.Length             2
#> 324   92  Petal.Width             1
#> 325   92 Sepal.Length             0
#> 326   92  Sepal.Width             2
#> 327   93 Petal.Length             1
#> 328   93  Petal.Width             0
#> 329   93 Sepal.Length             2
#> 330   93  Sepal.Width             3
#> 331   94 Petal.Length             0
#> 332   94  Petal.Width             1
#> 333   94  Sepal.Width             2
#> 334   95 Petal.Length             2
#> 335   95  Petal.Width             0
#> 336   95 Sepal.Length             3
#> 337   96 Petal.Length             0
#> 338   96  Petal.Width             3
#> 339   96  Sepal.Width             2
#> 340   97 Petal.Length             1
#> 341   97  Petal.Width             0
#> 342   97 Sepal.Length             2
#> 343   98 Petal.Length             4
#> 344   98  Petal.Width             0
#> 345   98 Sepal.Length             2
#> 346   99 Petal.Length             1
#> 347   99  Petal.Width             0
#> 348   99 Sepal.Length             2
#> 349   99  Sepal.Width             3
#> 350  100 Petal.Length             0
#> 351  100  Petal.Width             4
#> 352  100  Sepal.Width             3