Calculate minimal depth distribution of a random forest

Get minimal depth values for all trees in a random forest

min_depth_distribution(forest)

Arguments

forest
A randomForest or ranger object

Value

A data frame with the value of minimal depth for every variable in every tree

Examples

min_depth_distribution(randomForest::randomForest(Species ~ ., data = iris, ntree = 100))

#>     tree     variable minimal_depth
#> 1      1 Petal.Length             1
#> 2      1  Petal.Width             0
#> 3      1 Sepal.Length             4
#> 4      1  Sepal.Width             2
#> 5      2 Petal.Length             0
#> 6      2  Petal.Width             2
#> 7      2 Sepal.Length             2
#> 8      3 Petal.Length             2
#> 9      3  Petal.Width             0
#> 10     3  Sepal.Width             3
#> 11     4 Petal.Length             0
#> 12     4  Petal.Width             1
#> 13     4 Sepal.Length             2
#> 14     4  Sepal.Width             3
#> 15     5 Petal.Length             2
#> 16     5  Petal.Width             0
#> 17     5 Sepal.Length             1
#> 18     5  Sepal.Width             2
#> 19     6 Petal.Length             0
#> 20     6  Petal.Width             1
#> 21     6 Sepal.Length             3
#> 22     6  Sepal.Width             3
#> 23     7 Petal.Length             0
#> 24     7  Petal.Width             1
#> 25     7 Sepal.Length             2
#> 26     7  Sepal.Width             4
#> 27     8 Petal.Length             0
#> 28     8  Petal.Width             2
#> 29     8  Sepal.Width             3
#> 30     9 Petal.Length             5
#> 31     9  Petal.Width             0
#> 32     9 Sepal.Length             2
#> 33     9  Sepal.Width             3
#> 34    10 Petal.Length             0
#> 35    10  Petal.Width             2
#> 36    10 Sepal.Length             4
#> 37    10  Sepal.Width             3
#> 38    11 Petal.Length             1
#> 39    11  Petal.Width             0
#> 40    11 Sepal.Length             4
#> 41    11  Sepal.Width             2
#> 42    12 Petal.Length             2
#> 43    12  Petal.Width             0
#> 44    12 Sepal.Length             1
#> 45    12  Sepal.Width             4
#> 46    13 Petal.Length             2
#> 47    13  Petal.Width             0
#> 48    13 Sepal.Length             3
#> 49    13  Sepal.Width             2
#> 50    14 Petal.Length             0
#> 51    14  Petal.Width             1
#> 52    14  Sepal.Width             3
#> 53    15 Petal.Length             0
#> 54    15  Petal.Width             4
#> 55    15 Sepal.Length             4
#> 56    15  Sepal.Width             2
#> 57    16 Petal.Length             1
#> 58    16  Petal.Width             1
#> 59    16 Sepal.Length             0
#> 60    16  Sepal.Width             2
#> 61    17 Petal.Length             2
#> 62    17  Petal.Width             0
#> 63    18 Petal.Length             0
#> 64    18  Petal.Width             1
#> 65    18 Sepal.Length             3
#> 66    18  Sepal.Width             3
#> 67    19 Petal.Length             2
#> 68    19  Petal.Width             0
#> 69    20 Petal.Length             4
#> 70    20  Petal.Width             0
#> 71    20 Sepal.Length             2
#> 72    20  Sepal.Width             5
#> 73    21 Petal.Length             1
#> 74    21  Petal.Width             0
#> 75    21 Sepal.Length             3
#> 76    22 Petal.Length             0
#> 77    22  Petal.Width             3
#> 78    22 Sepal.Length             3
#> 79    22  Sepal.Width             3
#> 80    23 Petal.Length             0
#> 81    23  Petal.Width             1
#> 82    24 Petal.Length             3
#> 83    24  Petal.Width             0
#> 84    24 Sepal.Length             2
#> 85    24  Sepal.Width             4
#> 86    25 Petal.Length             0
#> 87    25  Petal.Width             1
#> 88    25 Sepal.Length             2
#> 89    25  Sepal.Width             3
#> 90    26 Petal.Length             1
#> 91    26  Petal.Width             2
#> 92    26 Sepal.Length             0
#> 93    26  Sepal.Width             1
#> 94    27 Petal.Length             1
#> 95    27  Petal.Width             0
#> 96    27 Sepal.Length             3
#> 97    27  Sepal.Width             4
#> 98    28 Petal.Length             0
#> 99    28  Petal.Width             2
#> 100   28  Sepal.Width             3
#> 101   29 Petal.Length             1
#> 102   29  Petal.Width             0
#> 103   29 Sepal.Length             2
#> 104   29  Sepal.Width             3
#> 105   30 Petal.Length             0
#> 106   30  Petal.Width             2
#> 107   30  Sepal.Width             3
#> 108   31 Petal.Length             2
#> 109   31  Petal.Width             0
#> 110   31 Sepal.Length             1
#> 111   32 Petal.Length             0
#> 112   32  Petal.Width             1
#> 113   32 Sepal.Length             3
#> 114   32  Sepal.Width             3
#> 115   33 Petal.Length             1
#> 116   33  Petal.Width             0
#> 117   33 Sepal.Length             2
#> 118   34 Petal.Length             0
#> 119   34  Petal.Width             2
#> 120   34 Sepal.Length             1
#> 121   34  Sepal.Width             3
#> 122   35 Petal.Length             2
#> 123   35  Petal.Width             1
#> 124   35 Sepal.Length             0
#> 125   35  Sepal.Width             3
#> 126   36 Petal.Length             2
#> 127   36  Petal.Width             0
#> 128   36 Sepal.Length             4
#> 129   36  Sepal.Width             2
#> 130   37 Petal.Length             1
#> 131   37  Petal.Width             0
#> 132   37 Sepal.Length             2
#> 133   37  Sepal.Width             6
#> 134   38 Petal.Length             2
#> 135   38  Petal.Width             0
#> 136   38  Sepal.Width             3
#> 137   39 Petal.Length             0
#> 138   39  Petal.Width             1
#> 139   39 Sepal.Length             3
#> 140   39  Sepal.Width             3
#> 141   40 Petal.Length             0
#> 142   40  Petal.Width             2
#> 143   40  Sepal.Width             2
#> 144   41 Petal.Length             1
#> 145   41  Petal.Width             2
#> 146   41 Sepal.Length             0
#> 147   41  Sepal.Width             5
#> 148   42 Petal.Length             2
#> 149   42  Petal.Width             0
#> 150   42 Sepal.Length             1
#> 151   43 Petal.Length             1
#> 152   43  Petal.Width             0
#> 153   43 Sepal.Length             3
#> 154   43  Sepal.Width             3
#> 155   44 Petal.Length             0
#> 156   44  Petal.Width             2
#> 157   44 Sepal.Length             2
#> 158   44  Sepal.Width             4
#> 159   45 Petal.Length             0
#> 160   45  Petal.Width             1
#> 161   46 Petal.Length             1
#> 162   46  Petal.Width             0
#> 163   46 Sepal.Length             1
#> 164   47 Petal.Length             0
#> 165   47 Sepal.Length             2
#> 166   47  Sepal.Width             3
#> 167   48 Petal.Length             0
#> 168   48  Petal.Width             3
#> 169   48 Sepal.Length             1
#> 170   48  Sepal.Width             4
#> 171   49 Petal.Length             2
#> 172   49  Petal.Width             0
#> 173   49 Sepal.Length             1
#> 174   49  Sepal.Width             2
#> 175   50 Petal.Length             1
#> 176   50  Petal.Width             0
#> 177   51 Petal.Length             0
#> 178   51  Petal.Width             2
#> 179   51 Sepal.Length             1
#> 180   51  Sepal.Width             3
#> 181   52 Petal.Length             0
#> 182   52  Petal.Width             3
#> 183   52 Sepal.Length             1
#> 184   52  Sepal.Width             2
#> 185   53 Petal.Length             1
#> 186   53  Petal.Width             0
#> 187   53  Sepal.Width             4
#> 188   54 Petal.Length             0
#> 189   54  Petal.Width             1
#> 190   54 Sepal.Length             3
#> 191   54  Sepal.Width             3
#> 192   55 Petal.Length             0
#> 193   55  Petal.Width             1
#> 194   55 Sepal.Length             2
#> 195   56 Petal.Length             0
#> 196   56  Petal.Width             1
#> 197   56 Sepal.Length             3
#> 198   56  Sepal.Width             4
#> 199   57 Petal.Length             1
#> 200   57  Petal.Width             1
#> 201   57 Sepal.Length             0
#> 202   57  Sepal.Width             3
#> 203   58 Petal.Length             1
#> 204   58  Petal.Width             1
#> 205   58 Sepal.Length             0
#> 206   59 Petal.Length             0
#> 207   59  Petal.Width             1
#> 208   59 Sepal.Length             4
#> 209   60 Petal.Length             2
#> 210   60  Petal.Width             0
#> 211   60  Sepal.Width             5
#> 212   61 Petal.Length             1
#> 213   61  Petal.Width             3
#> 214   61 Sepal.Length             1
#> 215   61  Sepal.Width             0
#> 216   62 Petal.Length             2
#> 217   62  Petal.Width             0
#> 218   62 Sepal.Length             3
#> 219   62  Sepal.Width             3
#> 220   63 Petal.Length             1
#> 221   63  Petal.Width             0
#> 222   63 Sepal.Length             2
#> 223   63  Sepal.Width             3
#> 224   64 Petal.Length             3
#> 225   64  Petal.Width             0
#> 226   64 Sepal.Length             1
#> 227   64  Sepal.Width             3
#> 228   65 Petal.Length             1
#> 229   65  Petal.Width             1
#> 230   65 Sepal.Length             0
#> 231   65  Sepal.Width             3
#> 232   66 Petal.Length             2
#> 233   66  Petal.Width             0
#> 234   66  Sepal.Width             5
#> 235   67 Petal.Length             0
#> 236   67 Sepal.Length             2
#> 237   68 Petal.Length             0
#> 238   68  Petal.Width             2
#> 239   68 Sepal.Length             3
#> 240   69 Petal.Length             0
#> 241   69  Petal.Width             1
#> 242   69 Sepal.Length             3
#> 243   69  Sepal.Width             4
#> 244   70 Petal.Length             0
#> 245   70  Petal.Width             2
#> 246   70  Sepal.Width             4
#> 247   71 Petal.Length             1
#> 248   71  Petal.Width             1
#> 249   71 Sepal.Length             0
#> 250   72 Petal.Length             1
#> 251   72  Petal.Width             0
#> 252   72 Sepal.Length             2
#> 253   72  Sepal.Width             3
#> 254   73 Petal.Length             0
#> 255   73  Petal.Width             1
#> 256   73 Sepal.Length             2
#> 257   73  Sepal.Width             4
#> 258   74 Petal.Length             0
#> 259   74  Petal.Width             1
#> 260   74 Sepal.Length             2
#> 261   75 Petal.Length             1
#> 262   75  Petal.Width             0
#> 263   75 Sepal.Length             3
#> 264   76 Petal.Length             1
#> 265   76  Petal.Width             0
#> 266   76  Sepal.Width             4
#> 267   77 Petal.Length             0
#> 268   77  Petal.Width             2
#> 269   77  Sepal.Width             3
#> 270   78 Petal.Length             3
#> 271   78  Petal.Width             1
#> 272   78 Sepal.Length             0
#> 273   78  Sepal.Width             1
#> 274   79 Petal.Length             3
#> 275   79  Petal.Width             0
#> 276   79 Sepal.Length             1
#> 277   79  Sepal.Width             4
#> 278   80 Petal.Length             1
#> 279   80  Petal.Width             1
#> 280   80 Sepal.Length             0
#> 281   80  Sepal.Width             4
#> 282   81 Petal.Length             2
#> 283   81  Petal.Width             0
#> 284   82 Petal.Length             2
#> 285   82  Petal.Width             0
#> 286   82 Sepal.Length             4
#> 287   82  Sepal.Width             3
#> 288   83 Petal.Length             1
#> 289   83  Petal.Width             1
#> 290   83 Sepal.Length             0
#> 291   84 Petal.Length             2
#> 292   84  Petal.Width             0
#> 293   84  Sepal.Width             2
#> 294   85 Petal.Length             0
#> 295   85  Petal.Width             2
#> 296   85 Sepal.Length             2
#> 297   85  Sepal.Width             3
#> 298   86 Petal.Length             1
#> 299   86  Petal.Width             2
#> 300   86 Sepal.Length             0
#> 301   86  Sepal.Width             1
#> 302   87 Petal.Length             0
#> 303   87  Petal.Width             3
#> 304   87  Sepal.Width             2
#> 305   88 Petal.Length             0
#> 306   88  Petal.Width             3
#> 307   88 Sepal.Length             2
#> 308   89 Petal.Length             0
#> 309   89  Petal.Width             2
#> 310   89 Sepal.Length             3
#> 311   89  Sepal.Width             3
#> 312   90 Petal.Length             1
#> 313   90  Petal.Width             1
#> 314   90 Sepal.Length             0
#> 315   90  Sepal.Width             4
#> 316   91 Petal.Length             0
#> 317   91  Petal.Width             4
#> 318   91 Sepal.Length             3
#> 319   91  Sepal.Width             3
#> 320   92 Petal.Length             0
#> 321   92  Petal.Width             2
#> 322   92  Sepal.Width             4
#> 323   93 Petal.Length             0
#> 324   93  Petal.Width             1
#> 325   93 Sepal.Length             2
#> 326   94 Petal.Length             0
#> 327   94  Petal.Width             3
#> 328   94 Sepal.Length             1
#> 329   94  Sepal.Width             4
#> 330   95 Petal.Length             2
#> 331   95  Petal.Width             0
#> 332   96 Petal.Length             1
#> 333   96  Petal.Width             0
#> 334   96 Sepal.Length             2
#> 335   96  Sepal.Width             3
#> 336   97 Petal.Length             1
#> 337   97  Petal.Width             0
#> 338   97 Sepal.Length             3
#> 339   97  Sepal.Width             2
#> 340   98 Petal.Length             0
#> 341   98  Petal.Width             1
#> 342   98  Sepal.Width             2
#> 343   99 Petal.Length             3
#> 344   99  Petal.Width             1
#> 345   99 Sepal.Length             0
#> 346  100 Petal.Length             1
#> 347  100  Petal.Width             0
#> 348  100 Sepal.Length             1
#> 349  100  Sepal.Width             2

min_depth_distribution(ranger::ranger(Species ~ ., data = iris, num.trees = 100))

#>     tree     variable minimal_depth
#> 1      1 Petal.Length             1
#> 2      1  Petal.Width             1
#> 3      1 Sepal.Length             0
#> 4      2 Petal.Length             0
#> 5      2  Petal.Width             1
#> 6      2 Sepal.Length             3
#> 7      3 Petal.Length             0
#> 8      3  Petal.Width             3
#> 9      3  Sepal.Width             3
#> 10     4 Petal.Length             0
#> 11     4  Petal.Width             2
#> 12     4 Sepal.Length             5
#> 13     4  Sepal.Width             3
#> 14     5 Petal.Length             3
#> 15     5  Petal.Width             0
#> 16     5 Sepal.Length             2
#> 17     6 Petal.Length             0
#> 18     6  Petal.Width             3
#> 19     6 Sepal.Length             1
#> 20     7 Petal.Length             0
#> 21     7  Petal.Width             2
#> 22     7 Sepal.Length             2
#> 23     8 Petal.Length             2
#> 24     8  Petal.Width             1
#> 25     8 Sepal.Length             0
#> 26     8  Sepal.Width             2
#> 27     9 Petal.Length             0
#> 28     9  Petal.Width             1
#> 29     9 Sepal.Length             2
#> 30     9  Sepal.Width             3
#> 31    10 Petal.Length             2
#> 32    10  Petal.Width             0
#> 33    10 Sepal.Length             4
#> 34    11 Petal.Length             0
#> 35    11  Petal.Width             1
#> 36    11 Sepal.Length             3
#> 37    11  Sepal.Width             2
#> 38    12 Petal.Length             1
#> 39    12  Petal.Width             0
#> 40    12 Sepal.Length             2
#> 41    13 Petal.Length             0
#> 42    13  Petal.Width             2
#> 43    13 Sepal.Length             1
#> 44    13  Sepal.Width             4
#> 45    14 Petal.Length             0
#> 46    14  Petal.Width             2
#> 47    14 Sepal.Length             1
#> 48    14  Sepal.Width             5
#> 49    15 Petal.Length             4
#> 50    15  Petal.Width             0
#> 51    15 Sepal.Length             3
#> 52    15  Sepal.Width             2
#> 53    16 Petal.Length             0
#> 54    16  Petal.Width             3
#> 55    16 Sepal.Length             1
#> 56    16  Sepal.Width             4
#> 57    17 Petal.Length             1
#> 58    17  Petal.Width             0
#> 59    17 Sepal.Length             2
#> 60    17  Sepal.Width             5
#> 61    18 Petal.Length             2
#> 62    18  Petal.Width             1
#> 63    18 Sepal.Length             0
#> 64    19 Petal.Length             0
#> 65    19  Petal.Width             1
#> 66    19 Sepal.Length             2
#> 67    20 Petal.Length             0
#> 68    20  Petal.Width             1
#> 69    20 Sepal.Length             4
#> 70    21 Petal.Length             0
#> 71    21  Petal.Width             3
#> 72    21 Sepal.Length             2
#> 73    21  Sepal.Width             3
#> 74    22 Petal.Length             1
#> 75    22  Petal.Width             1
#> 76    22 Sepal.Length             0
#> 77    22  Sepal.Width             2
#> 78    23 Petal.Length             0
#> 79    23  Petal.Width             1
#> 80    23 Sepal.Length             4
#> 81    23  Sepal.Width             3
#> 82    24 Petal.Length             2
#> 83    24  Petal.Width             0
#> 84    24 Sepal.Length             3
#> 85    25 Petal.Length             1
#> 86    25  Petal.Width             1
#> 87    25 Sepal.Length             0
#> 88    25  Sepal.Width             2
#> 89    26 Petal.Length             1
#> 90    26  Petal.Width             1
#> 91    26 Sepal.Length             0
#> 92    26  Sepal.Width             3
#> 93    27 Petal.Length             0
#> 94    27  Petal.Width             1
#> 95    27 Sepal.Length             2
#> 96    27  Sepal.Width             3
#> 97    28 Petal.Length             0
#> 98    28  Petal.Width             1
#> 99    28 Sepal.Length             2
#> 100   29 Petal.Length             1
#> 101   29  Petal.Width             0
#> 102   29 Sepal.Length             3
#> 103   30 Petal.Length             0
#> 104   30  Petal.Width             1
#> 105   30 Sepal.Length             2
#> 106   30  Sepal.Width             3
#> 107   31 Petal.Length             0
#> 108   31  Petal.Width             1
#> 109   31 Sepal.Length             4
#> 110   31  Sepal.Width             2
#> 111   32 Petal.Length             2
#> 112   32  Petal.Width             0
#> 113   32 Sepal.Length             1
#> 114   33 Petal.Length             1
#> 115   33  Petal.Width             0
#> 116   33 Sepal.Length             5
#> 117   33  Sepal.Width             3
#> 118   34 Petal.Length             0
#> 119   34  Petal.Width             1
#> 120   34  Sepal.Width             3
#> 121   35 Petal.Length             0
#> 122   35  Petal.Width             2
#> 123   35 Sepal.Length             3
#> 124   36 Petal.Length             2
#> 125   36  Petal.Width             0
#> 126   36 Sepal.Length             2
#> 127   36  Sepal.Width             5
#> 128   37 Petal.Length             2
#> 129   37  Petal.Width             0
#> 130   37  Sepal.Width             1
#> 131   38 Petal.Length             0
#> 132   38  Petal.Width             1
#> 133   38 Sepal.Length             2
#> 134   39 Petal.Length             2
#> 135   39  Petal.Width             1
#> 136   39 Sepal.Length             0
#> 137   39  Sepal.Width             1
#> 138   40 Petal.Length             0
#> 139   40  Petal.Width             2
#> 140   40 Sepal.Length             4
#> 141   41 Petal.Length             2
#> 142   41  Petal.Width             0
#> 143   42 Petal.Length             0
#> 144   42  Petal.Width             2
#> 145   42 Sepal.Length             1
#> 146   43 Petal.Length             0
#> 147   43  Petal.Width             2
#> 148   43 Sepal.Length             2
#> 149   44 Petal.Length             2
#> 150   44  Petal.Width             1
#> 151   44 Sepal.Length             0
#> 152   44  Sepal.Width             2
#> 153   45 Petal.Length             0
#> 154   45  Petal.Width             2
#> 155   45 Sepal.Length             1
#> 156   45  Sepal.Width             3
#> 157   46 Petal.Length             3
#> 158   46  Petal.Width             1
#> 159   46 Sepal.Length             0
#> 160   46  Sepal.Width             1
#> 161   47 Petal.Length             3
#> 162   47  Petal.Width             0
#> 163   47 Sepal.Length             3
#> 164   47  Sepal.Width             2
#> 165   48 Petal.Length             0
#> 166   48  Petal.Width             2
#> 167   48 Sepal.Length             1
#> 168   48  Sepal.Width             3
#> 169   49 Petal.Length             1
#> 170   49  Petal.Width             1
#> 171   49 Sepal.Length             0
#> 172   49  Sepal.Width             2
#> 173   50 Petal.Length             1
#> 174   50  Petal.Width             1
#> 175   50 Sepal.Length             0
#> 176   51 Petal.Length             2
#> 177   51  Petal.Width             0
#> 178   51 Sepal.Length             3
#> 179   51  Sepal.Width             2
#> 180   52 Petal.Length             2
#> 181   52  Petal.Width             1
#> 182   52 Sepal.Length             0
#> 183   53 Petal.Length             4
#> 184   53  Petal.Width             0
#> 185   53  Sepal.Width             3
#> 186   54 Petal.Length             2
#> 187   54  Petal.Width             1
#> 188   54 Sepal.Length             0
#> 189   54  Sepal.Width             1
#> 190   55 Petal.Length             3
#> 191   55  Petal.Width             0
#> 192   55 Sepal.Length             1
#> 193   55  Sepal.Width             4
#> 194   56 Petal.Length             0
#> 195   56  Petal.Width             1
#> 196   56 Sepal.Length             3
#> 197   56  Sepal.Width             3
#> 198   57 Petal.Length             0
#> 199   57 Sepal.Length             1
#> 200   58 Petal.Length             1
#> 201   58  Petal.Width             0
#> 202   59 Petal.Length             0
#> 203   59  Petal.Width             1
#> 204   59 Sepal.Length             3
#> 205   60 Petal.Length             2
#> 206   60  Petal.Width             1
#> 207   60 Sepal.Length             0
#> 208   60  Sepal.Width             3
#> 209   61 Petal.Length             1
#> 210   61  Petal.Width             2
#> 211   61 Sepal.Length             0
#> 212   61  Sepal.Width             4
#> 213   62 Petal.Length             0
#> 214   62  Petal.Width             3
#> 215   62 Sepal.Length             1
#> 216   62  Sepal.Width             4
#> 217   63 Petal.Length             0
#> 218   63  Petal.Width             2
#> 219   63 Sepal.Length             3
#> 220   63  Sepal.Width             5
#> 221   64 Petal.Length             0
#> 222   64  Petal.Width             2
#> 223   64 Sepal.Length             1
#> 224   64  Sepal.Width             4
#> 225   65 Petal.Length             0
#> 226   65  Petal.Width             1
#> 227   65  Sepal.Width             3
#> 228   66 Petal.Length             2
#> 229   66  Petal.Width             0
#> 230   66 Sepal.Length             1
#> 231   66  Sepal.Width             4
#> 232   67 Petal.Length             0
#> 233   67  Petal.Width             1
#> 234   67 Sepal.Length             3
#> 235   67  Sepal.Width             2
#> 236   68 Petal.Length             1
#> 237   68  Petal.Width             2
#> 238   68 Sepal.Length             0
#> 239   68  Sepal.Width             2
#> 240   69 Petal.Length             4
#> 241   69  Petal.Width             1
#> 242   69 Sepal.Length             0
#> 243   69  Sepal.Width             2
#> 244   70 Petal.Length             0
#> 245   70  Petal.Width             1
#> 246   70 Sepal.Length             2
#> 247   71 Petal.Length             1
#> 248   71  Petal.Width             0
#> 249   71 Sepal.Length             3
#> 250   71  Sepal.Width             3
#> 251   72  Petal.Width             1
#> 252   72 Sepal.Length             0
#> 253   72  Sepal.Width             3
#> 254   73 Petal.Length             6
#> 255   73  Petal.Width             0
#> 256   73 Sepal.Length             2
#> 257   73  Sepal.Width             3
#> 258   74 Petal.Length             1
#> 259   74  Petal.Width             0
#> 260   74 Sepal.Length             5
#> 261   75 Petal.Length             1
#> 262   75  Petal.Width             1
#> 263   75 Sepal.Length             0
#> 264   75  Sepal.Width             2
#> 265   76  Petal.Width             0
#> 266   76 Sepal.Length             1
#> 267   76  Sepal.Width             2
#> 268   77 Petal.Length             0
#> 269   77  Petal.Width             1
#> 270   77  Sepal.Width             3
#> 271   78 Petal.Length             1
#> 272   78 Sepal.Length             0
#> 273   79 Petal.Length             0
#> 274   79  Petal.Width             2
#> 275   79 Sepal.Length             2
#> 276   80 Petal.Length             0
#> 277   80  Petal.Width             1
#> 278   80 Sepal.Length             2
#> 279   81 Petal.Length             2
#> 280   81  Petal.Width             0
#> 281   81 Sepal.Length             1
#> 282   81  Sepal.Width             3
#> 283   82 Petal.Length             0
#> 284   82  Petal.Width             1
#> 285   82 Sepal.Length             2
#> 286   82  Sepal.Width             3
#> 287   83 Petal.Length             0
#> 288   83  Petal.Width             2
#> 289   83 Sepal.Length             1
#> 290   84 Petal.Length             1
#> 291   84  Petal.Width             0
#> 292   84  Sepal.Width             4
#> 293   85 Petal.Length             2
#> 294   85  Petal.Width             1
#> 295   85 Sepal.Length             0
#> 296   85  Sepal.Width             1
#> 297   86 Petal.Length             0
#> 298   86  Petal.Width             3
#> 299   86 Sepal.Length             2
#> 300   87 Petal.Length             0
#> 301   87  Petal.Width             1
#> 302   87 Sepal.Length             2
#> 303   87  Sepal.Width             3
#> 304   88 Petal.Length             1
#> 305   88  Petal.Width             2
#> 306   88 Sepal.Length             0
#> 307   88  Sepal.Width             1
#> 308   89 Petal.Length             0
#> 309   89  Petal.Width             4
#> 310   89 Sepal.Length             1
#> 311   89  Sepal.Width             2
#> 312   90 Petal.Length             2
#> 313   90  Petal.Width             0
#> 314   90 Sepal.Length             2
#> 315   90  Sepal.Width             3
#> 316   91 Petal.Length             0
#> 317   91  Petal.Width             1
#> 318   91 Sepal.Length             2
#> 319   92 Petal.Length             2
#> 320   92  Petal.Width             0
#> 321   92 Sepal.Length             1
#> 322   93 Petal.Length             0
#> 323   93  Petal.Width             2
#> 324   93 Sepal.Length             3
#> 325   94 Petal.Length             1
#> 326   94  Petal.Width             0
#> 327   94 Sepal.Length             2
#> 328   95 Petal.Length             3
#> 329   95  Petal.Width             0
#> 330   95 Sepal.Length             3
#> 331   96 Petal.Length             0
#> 332   96  Petal.Width             2
#> 333   96  Sepal.Width             5
#> 334   97 Petal.Length             0
#> 335   97  Petal.Width             1
#> 336   97 Sepal.Length             2
#> 337   97  Sepal.Width             4
#> 338   98 Petal.Length             0
#> 339   98  Petal.Width             1
#> 340   98 Sepal.Length             3
#> 341   99 Petal.Length             0
#> 342   99  Petal.Width             6
#> 343   99 Sepal.Length             2
#> 344   99  Sepal.Width             3
#> 345  100 Petal.Length             0
#> 346  100  Petal.Width             1
#> 347  100 Sepal.Length             4
#> 348  100  Sepal.Width             5